### 2023-03-29 #5311 @editorialbot
[PRE REVIEW]: PINA: Physics-Informed Neural networks for Advanced modeling
**Submitting author:** @ndem0 (<a href="http://orcid.org/0000-0003-3107-9738">Nicola Demo</a>)
**Repository:** https://github.com/mathLab/PINA
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.2
