### 2023-03-29 #5311 @danielskatz
👋 @ndem0 - Thanks for your submission. I'll now work on finding an editor for it.

In addition, you could work on the possibly missing and invalid DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
