+++ joss/paper.bib
@@ -2,6 +2,7 @@
   title={Deep learning: methods and applications},
   author={Deng, Li and Yu, Dong and others},
   journal={Foundations and trends{\textregistered} in signal processing},
+  doi = {10.1561/9781601988157},
   volume={7},
   number={3--4},
   pages={197--387},
@@ -30,7 +31,7 @@ volume = {378},
 pages = {686-707},
 year = {2019},
 issn = {0021-9991},
-doi = {https://doi.org/10.1016/j.jcp.2018.10.045},
+doi = {10.1016/j.jcp.2018.10.045},
 url = {https://www.sciencedirect.com/science/article/pii/S0021999118307125},
 author = {M. Raissi and P. Perdikaris and G.E. Karniadakis},
 keywords = {Data-driven scientific computing, Machine learning, Predictive modeling, Runge–Kutta methods, Nonlinear dynamics},
@@ -60,6 +61,7 @@ abstract = {We introduce physics-informed neural networks – neural networks th
   title={Neurodiffeq: A python package for solving differential equations with neural networks},
   author={Chen, Feiyu and Sondak, David and Protopapas, Pavlos and Mattheakis, Marios and Liu, Shuheng and Agarwal, Devansh and Di Giovanni, Marco},
   journal={Journal of Open Source Software},
+  doi = {10.21105/joss.01931},
   volume={5},
   number={46},
   pages={1931},
@@ -69,6 +71,7 @@ abstract = {We introduce physics-informed neural networks – neural networks th
   title={DeepXDE: A deep learning library for solving differential equations},
   author={Lu, Lu and Meng, Xuhui and Mao, Zhiping and Karniadakis, George Em},
   journal={SIAM Review},
+  doi = {10.1137/19m1274067},
   volume={63},
   number={1},
   pages={208--228},
@@ -91,6 +94,7 @@ abstract = {We introduce physics-informed neural networks – neural networks th
   title={NVIDIA SimNet™: An AI-accelerated multi-physics simulation framework},
   author={Hennigh, Oliver and Narasimhan, Susheela and Nabian, Mohammad Amin and Subramaniam, Akshay and Tangsali, Kaustubh and Fang, Zhiwei and Rietmann, Max and Byeon, Wonmin and Choudhry, Sanjay},
   booktitle={International Conference on Computational Science},
+  doi = {10.1007/978-3-030-77977-1_36},
   pages={447--461},
   year={2021},
   organization={Springer}
@@ -99,6 +103,7 @@ abstract = {We introduce physics-informed neural networks – neural networks th
   title={Sciann: A keras/tensorflow wrapper for scientific computations and physics-informed deep learning using artificial neural networks},
   author={Haghighat, Ehsan and Juanes, Ruben},
   journal={Computer Methods in Applied Mechanics and Engineering},
+  doi = {10.1016/j.cma.2020.113552},
   volume={373},
   pages={113552},
   year={2021},
@@ -124,7 +129,7 @@ volume = {360},
 pages = {112789},
 year = {2020},
 issn = {0045-7825},
-doi = {https://doi.org/10.1016/j.cma.2019.112789},
+doi = {10.1016/j.cma.2019.112789},
 url = {https://www.sciencedirect.com/science/article/pii/S0045782519306814},
 author = {Zhiping Mao and Ameya D. Jagtap and George Em Karniadakis},
 keywords = {Euler equations, Machine learning, Neural networks, Conservation laws, Riemann problem, Hidden fluid mechanics},
@@ -178,7 +183,7 @@ volume = {171},
 pages = {108875},
 year = {2022},
 issn = {0888-3270},
-doi = {https://doi.org/10.1016/j.ymssp.2022.108875},
[…]
