### 2023-04-17 #5352 @y-yao
One thing I'm wondering about in particular is: does the API for specifying boundary conditions allow for irregular boundaries (ie, non-rectangular). All examples seem to rely on rectangular boundaries and I don't see the constraints documented anywhere.
