### 2023-04-17 mathLab/PINA#78 @y-yao
API documentation
I didn't see any formal API documentaion page in the repo. For example, what are the assumptions and constraints of the various API functions.
One thing I'm wondering about in particular is: does the API for specifying boundary conditions allow for irregular boundaries (ie, non-rectangular). All examples seem to rely on rectangular boundaries and I don't see the constraints documented anywhere.

**Describe the solution you'd like**

**Additional context**
openjournals/joss-reviews#5352
