### 2023-04-17 #5352 @danielskatz
@y-yao - As you've discovered, `@editorialbot` is a fairly simple bot, only used to automate certain actions.  But your comments to the author about the part of your review you can't complete was great, as was your opening an issue about it.
