+++ docs/source/_rst/code.rst
@@ -9,3 +9,8 @@ Code Documentation
     DeepONet <deeponet.rst>
     PINN <pinn.rst>
     ContinuousConv <convolution.rst>
+    MultiFeedForward <multifeedforward.rst>
+    AbstractProblem <abstractproblem.rst>
+    SpatialProblem <spatialproblem.rst>
+    TimeDependentProblem <timedepproblem.rst>
+    ParametricProblem <parametricproblem.rst>
