+++ docs/source/_rst/timedepproblem.rst
@@ -1,0 +1,12 @@
+TimeDependentProblem
+===========
+.. currentmodule:: pina.problem.timedep_problem
+
+.. automodule:: pina.problem.timedep_problem
+
+.. autoclass:: TimeDependentProblem
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
+    :noindex:
