+++ docs/source/_rst/spatialproblem.rst
@@ -1,0 +1,12 @@
+SpatialProblem
+===========
+.. currentmodule:: pina.problem.spatial_problem
+
+.. automodule:: pina.problem.spatial_problem
+
+.. autoclass:: SpatialProblem
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
+    :noindex:
