+++ docs/source/_rst/parametricproblem.rst
@@ -1,0 +1,12 @@
+ParametricProblem
+===========
+.. currentmodule:: pina.problem.parametric_problem
+
+.. automodule:: pina.problem.parametric_problem
+
+.. autoclass:: ParametricProblem
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
+    :noindex:
