+++ docs/source/_rst/multifeedforward.rst
@@ -1,0 +1,12 @@
+MultiFeedForward
+===========
+.. currentmodule:: pina.model.multi_feed_forward
+
+.. automodule:: pina.model.multi_feed_forward
+
+.. autoclass:: MultiFeedForward
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
+    :noindex:
