+++ pina/span.py
@@ -7,8 +7,18 @@ from .utils import torch_lhs
 
 
 class Span(Location):
-    def __init__(self, span_dict):
+    """PINA implementation of Hypercube domain."""
 
+    def __init__(self, span_dict):
+        """
+        :param span_dict: A dictionary with dict-key a string representing
+            the input variables for the pinn, and dict-value a list with
+            the domain extrema.
+        :type span_dict: dict
+
+        :Example:
+            >>> spatial_domain = Span({'x': [0, 1], 'y': [0, 1]})
+        """
         self.fixed_ = {}
         self.range_ = {}
 
@@ -22,14 +32,46 @@ class Span(Location):
 
     @property
     def variables(self):
+        """Spatial variables.
+
+        :return: Spatial variables defined in '__init__()'
+        :rtype: list[str]
+        """
         return list(self.fixed_.keys()) + list(self.range_.keys())
 
     def update(self, new_span):
+        """Adding new dimensions on the span
+
+        :param new_span: A new span object to merge
+        :type new_span: Span
+
+        :Example:
+            >>> spatial_domain = Span({'x': [0, 1], 'y': [0, 1]})
+            >>> spatial_domain.variables
+            ['x', 'y']
+            >>> spatial_domain_2 = Span({'z': [3, 4], 'w': [0, 1]})
+            >>> spatial_domain.update(spatial_domain_2)
+            >>> spatial_domain.variables
+            ['x', 'y', 'z', 'w']
+        """
         self.fixed_.update(new_span.fixed_)
         self.range_.update(new_span.range_)
 
     def _sample_range(self, n, mode, bounds):
-        """
+        """Rescale the samples to the correct bounds
+
+        :param n: Number of points to sample, see Note below
+            for reference.
+        :type n: int
+        :param mode: Mode for sampling, defaults to 'random'.
+            Available modes include: random sampling, 'random';
+            latin hypercube sampling, 'latin' or 'lh';
+            chebyshev sampling, 'chebyshev'; grid sampling 'grid'.
+        :type mode: str, optional
+        :param bounds: Bounds to rescale the samples.
+        :type bounds: torch.Tensor
+        :return: Rescaled sample points.
+        :rtype: torch.Tensor
         """
         dim = bounds.shape[0]
         if mode in ['chebyshev', 'grid'] and dim != 1:
@@ -50,7 +92,55 @@ class Span(Location):
         return pts
 
     def sample(self, n, mode='random', variables='all'):
-        """TODO
+        """Sample routine.
+
+        :param n: Number of points to sample, see Note below
+            for reference.
+        :type n: int
+        :param mode: Mode for sampling, defaults to 'random'.
+            Available modes include: random sampling, 'random';
+            latin hypercube sampling, 'latin' or 'lh';
+            chebyshev sampling, 'chebyshev'; grid sampling 'grid'.
+        :type mode: str, optional
+        :param variables: pinn variable to be sampled, defaults to 'all'.
+        :type variables: str or list[str], optional
+
+        .. note::
+            The total number of points sampled in case of multiple variables
+            is not 'n', and it depends on the chosen 'mode'. If 'mode' is
+            'grid' or 'chebyshev', the points are sampled independentely
+            across the variables and the results crossed together, i.e. the
+            final number of points is 'n' to the power of the number of
+            variables. If 'mode' is 'random', 'lh' or 'latin', the variables
+            are sampled all together, and the final number of points
+
+        .. warning::
+            The extrema values of Span are always sampled only for 'grid' mode.
+
+        :Example:
+            >>> spatial_domain = Span({'x': [0, 1], 'y': [0, 1]})
[…]
