+++ docs/source/_rst/code.rst
@@ -4,8 +4,11 @@ Code Documentation
 .. toctree::
     :maxdepth: 3
 
+    PINN <pinn.rst>
     LabelTensor <label_tensor.rst>
+    Span <span.rst>
     FeedForward <fnn.rst>
     DeepONet <deeponet.rst>
-    PINN <pinn.rst>
     ContinuousConv <convolution.rst>
+    Operators <operators.rst>
+    Plotter <plotter.rst>
