### 2023-04-18 da33aea
partially fix documentation (#80)
