+++ pina/model/multi_feed_forward.py
@@ -1,13 +1,16 @@
+"""Module for Multi FeedForward model"""
 import torch
 
 from .feed_forward import FeedForward
 
 
 class MultiFeedForward(torch.nn.Module):
-
+    """
+    :param dict dff_dict: dictionary of FeedForward networks.
+    """
     def __init__(self, dff_dict):
-        '''
-        '''
+        """
+        """
         super().__init__()
 
         if not isinstance(dff_dict, dict):
