+++ docs/source/_rst/abstractproblem.rst
@@ -1,0 +1,12 @@
+AbstractProblem
+===========
+.. currentmodule:: pina.problem.abstract_problem
+
+.. automodule:: pina.problem.abstract_problem
+
+.. autoclass:: AbstractProblem
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
+    :noindex:
