+++ docs/source/_rst/code.rst
@@ -10,5 +10,9 @@ Code Documentation
     FeedForward <fnn.rst>
     DeepONet <deeponet.rst>
     ContinuousConv <convolution.rst>
+    MultiFeedForward <multifeedforward.rst>
+    AbstractProblem <abstractproblem.rst>
+    SpatialProblem <spatialproblem.rst>
+    TimeDependentProblem <timedepproblem.rst>
     Operators <operators.rst>
     Plotter <plotter.rst>
