+++ pina/location.py
@@ -5,10 +5,14 @@ from abc import ABCMeta, abstractmethod
 
 class Location(metaclass=ABCMeta):
     """
-    Abstract class
+    Abstract Location class.
+    Any geometry entity should inherit from this class.
     """
-
     @property
     @abstractmethod
     def sample(self):
+        """
+        Abstract method for sampling a point from the location. To be
+        implemented in the child class.
+        """
         pass
