+++ pina/problem/spatial_problem.py
@@ -16,22 +16,19 @@ class SpatialProblem(AbstractProblem):
         >>> from pina.operators import grad
         >>> from pina import Condition, Span
         >>> import torch
-
         >>> class SimpleODE(SpatialProblem):
-
         >>>     output_variables = ['u']
         >>>     spatial_domain = Span({'x': [0, 1]})
-
         >>>     def ode_equation(input_, output_):
         >>>         u_x = grad(output_, input_, components=['u'], d=['x'])
         >>>         u = output_.extract(['u'])
         >>>         return u_x - u
-
+        >>> 
         >>>     def initial_condition(input_, output_):
         >>>         value = 1.0
         >>>         u = output_.extract(['u'])
         >>>         return u - value
-
+        >>>
         >>>     conditions = {
         >>>         'x0': Condition(Span({'x': 0.}), initial_condition),
         >>>         'D': Condition(Span({'x': [0, 1]}), ode_equation)}
