+++ docs/source/_rst/multifeedforward.rst
@@ -1,5 +1,5 @@
 MultiFeedForward
-===========
+================
 .. currentmodule:: pina.model.multi_feed_forward
 
 .. automodule:: pina.model.multi_feed_forward
@@ -7,6 +7,4 @@ MultiFeedForward
 .. autoclass:: MultiFeedForward
     :members:
     :private-members:
-    :undoc-members:
     :show-inheritance:
-    :noindex:
