### 2023-04-18 2ca08b5
Docs (#81)
* clean `condition` module
* add docs
