### 2023-04-18 7182f35
minor updates (#82)
* new dev release
* github action architectures
