+++ pina/meta.py
@@ -14,7 +14,7 @@ __title__ = "pina"
 __author__ = "Nicola Demo, Maria Strazzullo"
 __copyright__ = "Copyright 2021-2021, PINA contributors"
 __license__ = "MIT"
-__version__ = "0.0.2"
+__version__ = "0.0.2.dev0"
 __mail__ = 'demo.nicola@gmail.com, '  # TODO
 __maintainer__ = __author__
 __status__ = "Alpha"
