+++ _sources/_rst/location.rst.txt
@@ -1,0 +1,10 @@
+Location
+=========
+.. currentmodule:: pina.location
+
+.. automodule:: pina.location
+
+.. autoclass:: Location
+    :members:
+    :private-members:
+    :show-inheritance:
