+++ _sources/_rst/operators.rst.txt
@@ -1,0 +1,10 @@
+Operators
+===========
+.. currentmodule:: pina.operators
+
+.. automodule:: pina.operators
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
+    :noindex:
