+++ _sources/_rst/code.rst.txt
@@ -4,8 +4,39 @@ Code Documentation
 .. toctree::
     :maxdepth: 3
 
+    PINN <pinn.rst>
     LabelTensor <label_tensor.rst>
+    Span <span.rst>
+    Operators <operators.rst>
+    Plotter <plotter.rst>
+    Condition <condition.rst>
+    Location <location.rst>
+
+Model
+-----
+
+.. toctree::
+    :maxdepth: 3
+
     FeedForward <fnn.rst>
     DeepONet <deeponet.rst>
-    PINN <pinn.rst>
+    MultiFeedForward <multifeedforward.rst>
+
+Layers
+------
+
+.. toctree::
+    :maxdepth: 3
+
     ContinuousConv <convolution.rst>
+
+Problem
+-------
+
+.. toctree::
+    :maxdepth: 3
+
+    AbstractProblem <abstractproblem.rst>
+    SpatialProblem <spatialproblem.rst>
+    TimeDependentProblem <timedepproblem.rst>
+    ParametricProblem <parametricproblem.rst>
