+++ _sources/_rst/abstractproblem.rst.txt
@@ -1,0 +1,11 @@
+AbstractProblem
+===============
+.. currentmodule:: pina.problem.abstract_problem
+
+.. automodule:: pina.problem.abstract_problem
+
+.. autoclass:: AbstractProblem
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
