+++ _sources/_rst/multifeedforward.rst.txt
@@ -1,0 +1,10 @@
+MultiFeedForward
+================
+.. currentmodule:: pina.model.multi_feed_forward
+
+.. automodule:: pina.model.multi_feed_forward
+
+.. autoclass:: MultiFeedForward
+    :members:
+    :private-members:
+    :show-inheritance:
