+++ _sources/_rst/timedepproblem.rst.txt
@@ -1,0 +1,11 @@
+TimeDependentProblem
+====================
+.. currentmodule:: pina.problem.timedep_problem
+
+.. automodule:: pina.problem.timedep_problem
+
+.. autoclass:: TimeDependentProblem
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
