+++ _sources/_rst/spatialproblem.rst.txt
@@ -1,0 +1,11 @@
+SpatialProblem
+==============
+.. currentmodule:: pina.problem.spatial_problem
+
+.. automodule:: pina.problem.spatial_problem
+
+.. autoclass:: SpatialProblem
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
