+++ _sources/_rst/condition.rst.txt
@@ -1,0 +1,10 @@
+Condition
+=========
+.. currentmodule:: pina.condition
+
+.. automodule:: pina.condition
+
+.. autoclass:: Condition
+    :members:
+    :private-members:
+    :show-inheritance:
