+++ _sources/_rst/parametricproblem.rst.txt
@@ -1,0 +1,10 @@
+ParametricProblem
+=================
+.. currentmodule:: pina.problem.parametric_problem
+
+.. automodule:: pina.problem.parametric_problem
+
+.. autoclass:: ParametricProblem
+    :members:
+    :private-members:
+    :show-inheritance:
