### 2023-04-24 mathLab/PINA#78 @ndem0
dear @y-yao, we are the documentation that was quite incomplete, now you can find it also online!
Regarding the possible shapes of the domain, the rectangular one is still the only public, but there are no limitations related to the software and/or the method. @dario-coscia will soon add the circular shape, but the user can implement its own inheriting from `Location`.
