+++ pina/ellipsoid.py
@@ -1,0 +1,231 @@
+import torch
+
+from .location import Location
+from .label_tensor import LabelTensor
+
+
+class Ellipsoid(Location):
+    """PINA implementation of Ellipsoid domain."""
+
+    def __init__(self, ellipsoid_dict, sample_surface=False):
+        """PINA implementation of Ellipsoid domain.
+
+        :param ellipsoid_dict: A dictionary with dict-key a string representing
+            the input variables for the pinn, and dict-value a list with
+            the domain extrema.
+        :type ellipsoid_dict: dict
+        :param sample_surface: A variable for choosing sample strategies. If
+            `sample_surface=True` only samples on the ellipsoid surface
+            frontier are taken. If `sample_surface=False` only samples on
+            the ellipsoid interior are taken, defaults to False.
+        :type sample_surface: bool, optional
+
+        .. warning::
+            Sampling for dimensions greater or equal to 10 could result
+            in a shrinking of the ellipsoid, which degrades the quality
+            of the samples. For dimensions higher than 10, other algorithms
+            for sampling should be used, such as: Dezert, Jean, and Christian
+            Musso. "An efficient method for generating points uniformly
+            distributed in hyperellipsoids." Proceedings of the Workshop on
+            Estimation, Tracking and Fusion: A Tribute to Yaakov Bar-Shalom.
+            Vol. 7. No. 8. 2001.
+
+        :Example:
+            >>> spatial_domain = Ellipsoid({'x':[1, 0], 'y':[0,1]})
+
+        """
+        self.fixed_ = {}
+        self.range_ = {}
+        self._centers = None
+        self._axis = None
+
+        if not isinstance(sample_surface, bool):
+            raise ValueError('sample_surface must be bool type.')
+
+        self._sample_surface = sample_surface
+
+        for k, v in ellipsoid_dict.items():
+            if isinstance(v, (int, float)):
+                self.fixed_[k] = v
+            elif isinstance(v, (list, tuple)) and len(v) == 2:
+                self.range_[k] = v
+            else:
+                raise TypeError
+
+        # perform operation only for not fixed variables (if any)
+
+        if self.range_:
+
+            # convert dict vals to torch [dim, 2] matrix
+            list_dict_vals = list(self.range_.values())
+            tmp = torch.tensor(list_dict_vals, dtype=torch.float)
+
+            # get the ellipsoid center
+            normal_basis = torch.eye(len(list_dict_vals))
+            centers = torch.diag(normal_basis * tmp.mean(axis=1))
+
+            # get the ellipsoid axis
+            ellipsoid_axis = (tmp - centers.reshape(-1, 1))[:, -1]
+
+            # save elipsoid axis and centers as dict
+            self._centers = dict(zip(self.range_.keys(), centers.tolist()))
+            self._axis = dict(zip(self.range_.keys(), ellipsoid_axis.tolist()))
+
+    @property
+    def variables(self):
+        """Spatial variables.
+
+        :return: Spatial variables defined in '__init__()'
+        :rtype: list[str]
+        """
+        return list(self.fixed_.keys()) + list(self.range_.keys())
+
+    def _sample_range(self, n, mode, variables):
+        """Rescale the samples to the correct bounds.
+
+        :param n: Number of points to sample in the ellipsoid.
+        :type n: int
+        :param mode: Mode for sampling, defaults to 'random'.
+            Available modes include: random sampling, 'random'.
+        :type mode: str, optional
+        :param variables: Variables to  be rescaled in the samples.
+        :type variables: torch.Tensor
+        :return: Rescaled sample points.
+        :rtype: torch.Tensor
+        """
+
+        # =============== For Developers ================ #
+        #
[…]
