+++ docs/source/_rst/ellipsoid.rst
@@ -1,0 +1,10 @@
+Ellipsoid
+===========
+.. currentmodule:: pina.ellipsoid
+
+.. automodule:: pina.ellipsoid
+
+.. autoclass:: Ellipsoid
+    :members:
+    :show-inheritance:
+    :noindex:
