### 2023-04-28 mathLab/PINA#78 @dario-coscia
hello! I have just added the `Ellipsoid` class in #86.
