### 2023-04-29 f6c1df2
is_inside method
