### 2023-05-02 b9dfbcd
test_extra_feats_pinn
