+++ tests/test_pinn.py
@@ -1,9 +1,9 @@
 import torch
 import pytest
 
-from pina import LabelTensor, Condition, Span, PINN
+from pina import LabelTensor, Condition, CartesianDomain, PINN
 from pina.problem import SpatialProblem
-from pina.model import FeedForward
+from pina.model import FeedForward, Network
 from pina.operators import nabla
 
 in_ = LabelTensor(torch.tensor([[0., 1.]]), ['x', 'y'])
@@ -11,7 +11,7 @@ out_ = LabelTensor(torch.tensor([[0.]]), ['u'])
 
 class Poisson(SpatialProblem):
     output_variables = ['u']
-    spatial_domain = Span({'x': [0, 1], 'y': [0, 1]})
+    spatial_domain = CartesianDomain({'x': [0, 1], 'y': [0, 1]})
 
     def laplace_equation(input_, output_):
         force_term = (torch.sin(input_.extract(['x'])*torch.pi) *
@@ -25,19 +25,19 @@ class Poisson(SpatialProblem):
 
     conditions = {
         'gamma1': Condition(
-            location=Span({'x': [0, 1], 'y':  1}),
+            location=CartesianDomain({'x': [0, 1], 'y':  1}),
             function=nil_dirichlet),
         'gamma2': Condition(
-            location=Span({'x': [0, 1], 'y': 0}),
+            location=CartesianDomain({'x': [0, 1], 'y': 0}),
             function=nil_dirichlet),
         'gamma3': Condition(
-            location=Span({'x':  1, 'y': [0, 1]}),
+            location=CartesianDomain({'x':  1, 'y': [0, 1]}),
             function=nil_dirichlet),
         'gamma4': Condition(
-            location=Span({'x': 0, 'y': [0, 1]}),
+            location=CartesianDomain({'x': 0, 'y': [0, 1]}),
             function=nil_dirichlet),
         'D': Condition(
-            location=Span({'x': [0, 1], 'y': [0, 1]}),
+            location=CartesianDomain({'x': [0, 1], 'y': [0, 1]}),
             function=laplace_equation),
         'data': Condition(
             input_points=in_,
@@ -53,13 +53,33 @@ class Poisson(SpatialProblem):
     truth_solution = poisson_sol
 
 
-problem = Poisson()
+class myFeature(torch.nn.Module):
+    """
+    Feature: sin(x)
+    """
 
-model = FeedForward(problem.input_variables, problem.output_variables)
+    def __init__(self):
+        super(myFeature, self).__init__()
+
+
+    def forward(self, x):
+        t = (torch.sin(x.extract(['x'])*torch.pi) *
+             torch.sin(x.extract(['y'])*torch.pi))
+        return LabelTensor(t, ['sin(x)sin(y)'])
+
+
+problem = Poisson()
+model = FeedForward(len(problem.input_variables),len(problem.output_variables))
+model_extra_feat =  FeedForward(len(problem.input_variables) + 1,len(problem.output_variables))
+
 
 
 def test_constructor():
     PINN(problem, model)
+
+
+def test_constructor_extra_feats():
+    PINN(problem, model_extra_feat, [myFeature()])
 
 
 def test_span_pts():
@@ -127,6 +147,28 @@ def test_train_2():
     param = [0, 3]
     for i, truth_key in zip(param, expected_keys):
         pinn = PINN(problem, model)
+        pinn.span_pts(n, 'grid', locations=boundaries)
+        pinn.span_pts(n, 'grid', locations=['D'])
+        pinn.train(50, save_loss=i)
+        assert list(pinn.history_loss.keys()) == truth_key
+
+
[…]
