+++ pina/__init__.py
@@ -3,7 +3,7 @@ __all__ = [
     'LabelTensor',
     'Plotter',
     'Condition',
-    'Span',
+    'CartesianDomain',
     'Location',
 ]
 
@@ -11,6 +11,6 @@ from .meta import *
 from .label_tensor import LabelTensor
 from .pinn import PINN
 from .plotter import Plotter
-from .span import Span
+from .cartesian import CartesianDomain
 from .condition import Condition
 from .location import Location
