### 2023-05-03 4cc1b88
equation class, fix minor bugs, diff domain
