+++ pina/geometry/__init__.py
@@ -1,0 +1,10 @@
+__all__ = [
+    'Location',
+    'CartesianDomain',
+    'EllypsoidDomain',
+]
+
+from .location import Location
+from .cartesian import CartesianDomain
+from .ellypsoid import EllypsoidDomain
+from .difference_domain import Difference
