+++ pina/model/network.py
@@ -59,10 +59,11 @@ class Network(torch.nn.Module):
         :raises ValueError: Error in constructing the PINA network
         """
         try:
-            tmp = torch.rand((10, len(self._input_variables)))
-            tmp = LabelTensor(tmp, self._input_variables)
-            tmp = self.forward(tmp)  # trying a forward pass
-            tmp = LabelTensor(tmp, self._output_variables)
+            pass
+            # tmp = torch.rand((10, len(self._input_variables)))
+            # tmp = LabelTensor(tmp, self._input_variables)
+            # tmp = self.forward(tmp)  # trying a forward pass
+            # tmp = LabelTensor(tmp, self._output_variables)
         except:
             raise ValueError('Error in constructing the PINA network.'
                              ' Check compatibility of input/output'
