+++ docs/source/_rst/tutorial1/tutorial.rst
@@ -82,11 +82,12 @@ time domain where we want the solution.
 Summarizing, in PINA we can initialize a problem with a class which is
 inherited from three base classes: ``SpatialProblem``,
 ``TimeDependentProblem``, ``ParametricProblem``, depending on the type
-of problem we are considering. For reference:
-
-* ``SpatialProblem`` :math:`\rightarrow` spatial variable(s) presented in the differential equation 
-* ``TimeDependentProblem`` :math:`\rightarrow` time variable(s) presented in the differential equation 
-* ``ParametricProblem`` :math:`\rightarrow` parameter(s) presented in the differential equation
+of problem we are considering. For reference: \* ``SpatialProblem``
+:math:`\rightarrow` spatial variable(s) presented in the differential
+equation \* ``TimeDependentProblem`` :math:`\rightarrow` time
+variable(s) presented in the differential equation \*
+``ParametricProblem`` :math:`\rightarrow` parameter(s) presented in the
+differential equation
 
 Write the problem class
 ~~~~~~~~~~~~~~~~~~~~~~~
@@ -136,8 +137,8 @@ Equation (1) and try to write the PINA model class:
     
         # Conditions to hold
         conditions = {
-            'x0': Condition(Span({'x': 0.}), initial_condition),
-            'D': Condition(Span({'x': [0, 1]}), ode_equation),
+            'x0': Condition(location=Span({'x': 0.}), function=initial_condition),
+            'D': Condition(location=Span({'x': [0, 1]}), function=ode_equation),
         }
     
         # defining true solution
@@ -261,6 +262,7 @@ the results.
                   sum          x0initial_co Dode_equatio 
     [epoch 02000] 1.099083e-03 3.420736e-05 1.064875e-03 
     [epoch 03000] 4.049759e-04 2.937766e-06 4.020381e-04 
+
 
 
 After the training we have saved the final loss in ``final_loss``, which
