### 2023-05-15 5ba5200
plotter.py plot_loss fixed Issue #95
