### 2023-05-15 95937a5
dummy dataloader, writer, refactoring
