+++ pina/pinn.py
@@ -5,7 +5,8 @@ import torch.optim.lr_scheduler as lrs
 from .problem import AbstractProblem
 from .model import Network
 from .label_tensor import LabelTensor
-from .utils import merge_tensors, PinaDataset
+from .utils import merge_tensors
+from .dataset import DummyLoader
 
 
 torch.pi = torch.acos(torch.zeros(1)).item() * 2  # which is 3.1415927410125732
@@ -26,6 +27,7 @@ class PINN(object):
                  batch_size=None,
                  dtype=torch.float32,
                  device='cpu',
+                 writer=None,
                  error_norm='mse'):
         '''
         :param AbstractProblem problem: the formualation of the problem.
@@ -84,17 +86,22 @@ class PINN(object):
         self.input_pts = {}
 
         self.trained_epoch = 0
+    
+        from .writer import Writer
+        if writer is None:
+            writer = Writer()
+        self.writer = writer
 
         if not optimizer_kwargs:
             optimizer_kwargs = {}
         optimizer_kwargs['lr'] = lr
         self.optimizer = optimizer(
-            self.model.parameters(), weight_decay=regularizer, **optimizer_kwargs)
-        self._lr_scheduler = lr_scheduler_type(
-            self.optimizer, **lr_scheduler_kwargs)
+            self.model.parameters())#, weight_decay=regularizer, **optimizer_kwargs)
+        #self._lr_scheduler = lr_scheduler_type(
+        #    self.optimizer, **lr_scheduler_kwargs)
 
         self.batch_size = batch_size
-        self.data_set = PinaDataset(self)
+        # self.data_set = PinaDataset(self)
 
     @property
     def problem(self):
@@ -241,44 +248,44 @@ class PINN(object):
         """
         input_pts = input_pts.to(dtype=self.dtype, device=self.device)
         output_pts = output_pts.to(dtype=self.dtype, device=self.device)
-        predicted = self.model(pts)
+        predicted = self.model(input_pts)
         residuals = predicted - output_pts
         return self._compute_norm(residuals)
  
 
-    def closure(self):
-        """
-        """
-        self.optimizer.zero_grad()
+    # def closure(self):
+    #     """
+    #     """
+    #     self.optimizer.zero_grad()
 
-        condition_losses = []
-        from torch.utils.data import DataLoader
-        from .utils import MyDataset
-        loader = DataLoader(
-            MyDataset(self.input_pts),
-            batch_size=self.batch_size,
-            num_workers=1
-        )
-        for condition_name in self.problem.conditions:
-            condition = self.problem.conditions[condition_name]
+    #     condition_losses = []
+    #     from torch.utils.data import DataLoader
+    #     from .utils import MyDataset
+    #     loader = DataLoader(
+    #         MyDataset(self.input_pts),
+    #         batch_size=self.batch_size,
+    #         num_workers=1
+    #     )
+    #     for condition_name in self.problem.conditions:
+    #         condition = self.problem.conditions[condition_name]
 
-            batch_losses = []
-            for batch in data_loader[condition_name]:
+    #         batch_losses = []
+    #         for batch in data_loader[condition_name]:
 
-                if hasattr(condition, 'equation'):
-                    loss = self._residual_loss(
-                        batch[condition_name], condition.equation)
-                elif hasattr(condition, 'output_points'):
-                    loss = self._data_loss(
-                        batch[condition_name], condition.output_points)
+    #             if hasattr(condition, 'equation'):
+    #                 loss = self._residual_loss(
+    #                     batch[condition_name], condition.equation)
+    #             elif hasattr(condition, 'output_points'):
+    #                 loss = self._data_loss(
+    #                     batch[condition_name], condition.output_points)
 
[…]

[…]
