+++ pina/operators.py
@@ -65,6 +65,7 @@ def grad(output_, input_, components=None, d=None):
         return gradients
 
     if not isinstance(input_, LabelTensor):
+        print(input_.type())
         raise TypeError
 
     if d is None:
