### 2023-05-15 2827e9f
added legend support for plot_loss and some comments
