+++ pina/pinn.py
@@ -5,7 +5,8 @@ import torch.optim.lr_scheduler as lrs
 from .problem import AbstractProblem
 from .model import Network
 from .label_tensor import LabelTensor
-from .utils import merge_tensors, PinaDataset
+from .utils import merge_tensors
+from .dataset import DummyLoader
 
 
 torch.pi = torch.acos(torch.zeros(1)).item() * 2  # which is 3.1415927410125732
@@ -26,6 +27,7 @@ class PINN(object):
                  batch_size=None,
                  dtype=torch.float32,
                  device='cpu',
+                 writer=None,
                  error_norm='mse'):
         '''
         :param AbstractProblem problem: the formualation of the problem.
@@ -84,17 +86,22 @@ class PINN(object):
         self.input_pts = {}
 
         self.trained_epoch = 0
+    
+        from .writer import Writer
+        if writer is None:
+            writer = Writer()
+        self.writer = writer
 
         if not optimizer_kwargs:
             optimizer_kwargs = {}
         optimizer_kwargs['lr'] = lr
         self.optimizer = optimizer(
-            self.model.parameters(), weight_decay=regularizer, **optimizer_kwargs)
-        self._lr_scheduler = lr_scheduler_type(
-            self.optimizer, **lr_scheduler_kwargs)
+            self.model.parameters())#, weight_decay=regularizer, **optimizer_kwargs)
+        #self._lr_scheduler = lr_scheduler_type(
+        #    self.optimizer, **lr_scheduler_kwargs)
 
         self.batch_size = batch_size
-        self.data_set = PinaDataset(self)
+        # self.data_set = PinaDataset(self)
 
     @property
     def problem(self):
@@ -216,139 +223,131 @@ class PINN(object):
             # pts = pts.double()
             self.input_pts[location] = pts
 
-    def train(self, stop=100, frequency_print=2, save_loss=1, trial=None):
+    def _residual_loss(self, input_pts, equation):
+        """
+        Compute the residual loss for a given condition.
 
-        self.model.train()
-        epoch = 0
-        # Add all condition with `input_points` to dataloader
-        for condition in list(set(self.problem.conditions.keys()) - set(self.input_pts.keys())):
-            self.input_pts[condition] = self.problem.conditions[condition]
+        :param torch.Tensor pts: the points to evaluate the residual at.
+        :param Equation equation: the equation to evaluate the residual with.
+        """
 
-        data_loader = self.data_set.dataloader
+        input_pts = input_pts.to(dtype=self.dtype, device=self.device)
+        input_pts.requires_grad_(True)
+        input_pts.retain_grad()
 
-        header = []
-        for condition_name in self.problem.conditions:
-            condition = self.problem.conditions[condition_name]
+        predicted = self.model(input_pts)
+        residuals = equation.residual(input_pts, predicted)
+        return self._compute_norm(residuals)
 
-            if hasattr(condition, 'function'):
-                if isinstance(condition.function, list):
-                    for function in condition.function:
-                        header.append(f'{condition_name}{function.__name__}')
+    def _data_loss(self, input_pts, output_pts):
+        """
+        Compute the residual loss for a given condition.
 
+        :param torch.Tensor pts: the points to evaluate the residual at.
+        :param Equation equation: the equation to evaluate the residual with.
+        """
+        input_pts = input_pts.to(dtype=self.dtype, device=self.device)
+        output_pts = output_pts.to(dtype=self.dtype, device=self.device)
+        predicted = self.model(input_pts)
+        residuals = predicted - output_pts
+        return self._compute_norm(residuals)
+ 
+
+    # def closure(self):
+    #     """
+    #     """
+    #     self.optimizer.zero_grad()
+
[…]
