+++ pina/geometry/__init__.py
@@ -1,0 +1,10 @@
+__all__ = [
+    'Location',
+    'CartesianDomain',
+    'EllipsoidDomain',
+]
+
+from .location import Location
+from .cartesian import CartesianDomain
+from .ellipsoid import EllipsoidDomain
+from .difference_domain import Difference
