### 2023-05-15 9b984bc
Improved the plot_loss to show title and labels. Added support for saving the plot to a file.
