### 2023-05-15 0c10218
Allows to plot in a single plot.
