### 2023-05-19 mathLab/PINA#102 @yorkiva
Adding Regularization to the PINN Loss function
The condition module is the only place where you can define your physics-informed loss functions. However, there seems to be no way of adding additional regularizers like a L1 or L2 regularizer on the model's weights. Such regularization can be very useful to constrain overfitting, especially if training with noisy IC.

**Describe the solution you'd like**
It would be good to have an extension of the condition module (or something similar) to allow regularization.

**Additional context**
https://github.com/openjournals/joss-reviews/issues/5352
