### 2023-05-19 mathLab/PINA#103 @yorkiva
Adding Comments to the Examples
The package contains a good set of examples. But their comprehensibility is not up to the mark, especially for someone who is just starting with the package and not yet familiar with the different modules and their utilities.

**Describe the solution you'd like**
In the examples, comments should be added to make it clear what problem is being solved, what are the domains, constraints etc. Additional comments should be added to clearly illustrate what different parts of the code are accomplishing.

**Additional context**
https://github.com/openjournals/joss-reviews/issues/5352
