### 2023-05-19 mathLab/PINA#103 @yorkiva
In fact, I think it would be superb to convert some of the examples into notebooks. Some of these examples seem small, even possibly trainable on CPUs. It would be good to have notebooks that clearly show the logging during the progress, the kind of outputs it generates, and how to do inferencing and make plots.
