### 2023-05-20 mathLab/PINA#103 @dario-coscia
👋🏻 @yorkiva Thanks for the feedback! Indeed you are right, we will add comments to the examples. Some of the interesting problems (e.g Stokes) are a bit too long to train on a notebook, but we have a list of possible tutorials [here](https://github.com/mathLab/PINA/tree/master/tutorials) in `.py`, `.html`, `.ipynb` formats. These tutorials comprehend different types of problems (e.g. time dependent or spatial problems). We will add a tutorial for parametric dependent problem, maybe following the Poisson example in `example` folder.
