### 2023-05-25 mathLab/PINA#102 @dario-coscia
👋🏻 @yorkiva Thank you for your comment. The `Condition` class is used to define the condition to apply (e.g. function) and where to apply it (e.g. location). The L2 regulariser on the weights can be accomplished by setting the `regularizer` parameter in the `PINN` class.

Eventually, I agree with you that a `
class `Loss`  should be implemented, to enable the user to make this changes. We will release soon a beta version of the software where we plan to introduce these features, but for mantainability we can not merge these features on the current version.
