+++ pina/model/__init__.py
@@ -2,10 +2,8 @@ __all__ = [
     'FeedForward',
     'MultiFeedForward',
     'DeepONet',
-    'Network',
 ]
 
 from .feed_forward import FeedForward
 from .multi_feed_forward import MultiFeedForward
 from .deeponet import DeepONet
-from .network import Network

[…]
