### 2023-05-29 2d56434
minor changes/ loss class
