+++ tests/test_pinn.py
@@ -1,16 +1,16 @@
 import torch
 import pytest
 
-from pina import LabelTensor, Condition, CartesianDomain, PINN
 from pina.problem import SpatialProblem
-from pina.model import FeedForward
 from pina.operators import nabla
+from pina.geometry import CartesianDomain
+from pina import Condition, LabelTensor, PINN
+from pina.trainer import Trainer
+from pina.model import FeedForward
 from pina.equation.equation import Equation
 from pina.equation.equation_factory import FixedValue
+from pina.plotter import Plotter
 
-
-in_ = LabelTensor(torch.tensor([[0., 1.]]), ['x', 'y'])
-out_ = LabelTensor(torch.tensor([[0.]]), ['u'])
 
 def laplace_equation(input_, output_):
     force_term = (torch.sin(input_.extract(['x'])*torch.pi) *
@@ -19,6 +19,8 @@ def laplace_equation(input_, output_):
     return nabla_u - force_term
 
 my_laplace = Equation(laplace_equation)
+in_ = LabelTensor(torch.tensor([[0., 1.]], requires_grad=True), ['x', 'y'])
+out_ = LabelTensor(torch.tensor([[0.]], requires_grad=True), ['u'])
 
 class Poisson(SpatialProblem):
     output_variables = ['u']
@@ -68,75 +70,75 @@ class myFeature(torch.nn.Module):
         return LabelTensor(t, ['sin(x)sin(y)'])
 
 
-problem = Poisson()
-model = FeedForward(len(problem.input_variables),len(problem.output_variables))
-model_extra_feat =  FeedForward(len(problem.input_variables) + 1,len(problem.output_variables))
+# make the problem
+poisson_problem = Poisson()
+model = FeedForward(len(poisson_problem.input_variables),len(poisson_problem.output_variables))
+model_extra_feats = FeedForward(len(poisson_problem.input_variables)+1,len(poisson_problem.output_variables))
+extra_feats = [myFeature()]
 
 
 def test_constructor():
-    PINN(problem, model)
+    PINN(problem = poisson_problem, model=model, extra_features=None)
 
 
 def test_constructor_extra_feats():
-    PINN(problem, model_extra_feat, [myFeature()])
+    model = FeedForward(len(poisson_problem.input_variables)+1,len(poisson_problem.output_variables))
+    PINN(problem = poisson_problem, model=model_extra_feats, extra_features=extra_feats)
 
-
-def test_span_pts():
-    pinn = PINN(problem, model)
+# TODO move in test_problem
+def test_discretise_domain():
     n = 10
     boundaries = ['gamma1', 'gamma2', 'gamma3', 'gamma4']
-    pinn.span_pts(n, 'grid', locations=boundaries)
+    poisson_problem.discretise_domain(n, 'grid', locations=boundaries)
     for b in boundaries:
-        assert pinn.input_pts[b].shape[0] == n
-    pinn.span_pts(n, 'random', locations=boundaries)
+        assert poisson_problem.input_pts[b].shape[0] == n
+    poisson_problem.discretise_domain(n, 'random', locations=boundaries)
     for b in boundaries:
-        assert pinn.input_pts[b].shape[0] == n
+        assert poisson_problem.input_pts[b].shape[0] == n
 
-    pinn.span_pts(n, 'grid', locations=['D'])
-    assert pinn.input_pts['D'].shape[0] == n**2
-    pinn.span_pts(n, 'random', locations=['D'])
-    assert pinn.input_pts['D'].shape[0] == n
[…]
