### 2023-05-29 mathLab/PINA#103 @dario-coscia
👋🏻 @yorkiva . I updated the example folder (see https://github.com/mathLab/PINA/pull/105), so now every problem in `example/problem` is documented. Few of the problems are not available at the moments.

We will soon release a beta version of PINA where some high level features will change (rename some classes, and add more flexibility for the user by enabling pytorch lightining). Since updating the tutorial requires a big effort (for documentation compatibility, etc.), and due to the soon release of the beta verison, we will not make a tutorial on Stokes (but you can see the working example in `example` folder`).

Anyway, thank you for the feedback, we will also take into account when working on the beta version #85 .
