### 2023-05-29 mathLab/PINA#102 @dario-coscia
Hello @yorkiva :)

Just wanted to let you know that in the beta version we will soon release the possibility to add a custom loss for the PINN (https://github.com/mathLab/PINA/pull/105), and other very cool features such as gradient clipping, batch gradient accumulation, ... since we will use lightining `Trainer` module in backhand to train the PINN .

Thank you for the very useful feedbacks 😄
