### 2023-05-29 1d88334
data pinn  loss added
