+++ pina/problem/abstract_problem.py
@@ -14,9 +14,17 @@ class AbstractProblem(metaclass=ABCMeta):
     """
 
     def __init__(self):
-        self.have_sampled_points = False
+
+        # variable storing all points
         self.input_pts = {}
-        
+
+        # varible to check if sampling is done. If no location
+        # element is presented in Condition this variable is set to true
+        self.have_sampled_points = True
+
+        # put in self.input_pts all the points that we don't need to sample
+        self._span_condition_points()
+      
     @property
     def input_variables(self):
         """
@@ -87,6 +95,22 @@ class AbstractProblem(metaclass=ABCMeta):
         """
         pass
 
+    def _span_condition_points(self):
+        """
+        Simple function to get the condition points
+        """
+        for condition_name in self.conditions:
+            condition = self.conditions[condition_name]
+            if hasattr(condition, 'equation') and hasattr(condition, 'input_points'):
+                samples = condition.input_points
+            elif hasattr(condition, 'output_points') and hasattr(condition, 'input_points'):
+                samples = (condition.input_points, condition.output_points)
+            # skip if we need to sample
+            elif hasattr(condition, 'location'):
+                self.have_sampled_points = False
+                continue
+            self.input_pts[condition_name] = samples
+
     def discretise_domain(self, *args, **kwargs):
         """
         Generate a set of points to span the `Location` of all the conditions of
@@ -96,7 +120,6 @@ class AbstractProblem(metaclass=ABCMeta):
         >>> pinn.span_pts(n=10, mode='grid', location=['bound1'])
         >>> pinn.span_pts(n=10, mode='grid', variables=['x'])
         """
-
         if all(key in kwargs for key in ['n', 'mode']):
             argument = {}
             argument['n'] = kwargs['n']
@@ -130,9 +153,9 @@ class AbstractProblem(metaclass=ABCMeta):
                             for argument in arguments)
             pts = merge_tensors(samples)
             self.input_pts[location] = pts
-
             # setting the grad
             self.input_pts[location].requires_grad_(True)
             self.input_pts[location].retain_grad()
-        
+
+        # all points have been sampled
         self.have_sampled_points = True
