### 2023-06-01 ec6ea99
Remove plt.show() in case of savefig
