### 2023-06-01 mathLab/PINA#107 @akshaysubr
Tutorial in the documentation fails when run
**Describe the bug**
The first tutorial example fails because the `Condition` arguments are wrong
https://mathlab.github.io/PINA/_rst/tutorial1/tutorial.html

**To Reproduce**
The code in https://mathlab.github.io/PINA/_rst/tutorial1/tutorial.html

**Expected behavior**
The example should run fine

**Output**
```
Traceback (most recent call last):
  File "example.py", line 8, in <module>
    class SimpleODE(SpatialProblem):
  File "example.py", line 39, in SimpleODE
    'x0': Condition(Span({'x': 0.}), initial_condition),
  File "/code/pina/condition.py", line 69, in __init__
    raise ValueError('Condition takes only the following keyword arguments: {`input_points`, `output_points`, `location`, `function`, `data_weight`}.')
ValueError: Condition takes only the following keyword arguments: {`input_points`, `output_points`, `location`, `function`, `data_weight`}.
```

**Additional context**
https://github.com/openjournals/joss-reviews/issues/5352
