### 2023-06-02 mathLab/PINA#107 @dario-coscia
👋🏻  @akshaysubr and thanks for the feedback. I checked the tutorial in `.pynb`, `.py` in [here](https://github.com/mathLab/PINA/tree/master/tutorials/tutorial1), and the `.rst` file which is used when building the docs in [here](https://github.com/mathLab/PINA/tree/master/docs/source/_rst/tutorial1) and they are all correctly updated.

The problem you are encountering is due to ReadTheDocs update. Indeed, we update automatically the documentation page once a month, so if we make changes they are not immediately updated on the documentation page but only once a month. This avoids to refresh the doc page after every commit. Recently commit #91, we made a modification in the tutorial due to a modification in the `Conditioning` class, so this is not yet visible in the doc page.

Hope this helps :)
