+++ _sources/_rst/network.rst.txt
@@ -1,0 +1,10 @@
+Network
+================
+.. currentmodule:: pina.model.network
+
+.. automodule:: pina.model.network
+
+.. autoclass:: Network
+    :members:
+    :private-members:
+    :show-inheritance:
