### 2023-06-02 mathLab/PINA#107 @akshaysubr
Thanks, that helps. Confirming that the `.rst` file has the right updated documentation:
```
    conditions = {
        'x0': Condition(location=Span({'x': 0.}), function=initial_condition),
        'D': Condition(location=Span({'x': [0, 1]}), function=ode_equation),
    }
```

Would be good to have the readthedocs page also updated or maybe mention this one month lag in the readme so in case something doesn't work, people can check out the native docs in the repo.
