### 2023-06-03 mathLab/PINA#107 @dario-coscia
Hello! We updated the documentation page, now also there the tutorial is updated.
Thanks for the feedback, we will discuss how to put this into the README😊
