+++ pina/label_tensor.py
@@ -53,7 +53,7 @@ class LabelTensor(torch.Tensor):
                     [0.8207, 0.9239],
                     [0.5545, 0.8927],
                     ...,
-                    [0.9427, 0.5819]
+                    [0.9427, 0.5819],
                     [0.9518, 0.1025],
                     [0.8066, 0.9615]])
         '''
@@ -89,6 +89,7 @@ class LabelTensor(torch.Tensor):
         self._labels = labels   # assign the label
 
     # TODO remove try/ except thing IMPORTANT
+    # make the label None of default
     def clone(self, *args, **kwargs):
         """
         Clone the LabelTensor. For more details, see
