+++ pina/__init__.py
@@ -1,5 +1,6 @@
 __all__ = [
     'PINN',
+    'Trainer',
     'LabelTensor',
     'Plotter',
     'Condition',
@@ -10,6 +11,7 @@ __all__ = [
 from .meta import *
 from .label_tensor import LabelTensor
 from .pinn import PINN
+from .trainer import Trainer
 from .plotter import Plotter
 from .condition import Condition
 from .geometry import Location
