+++ pina/label_tensor.py
@@ -88,6 +88,8 @@ class LabelTensor(torch.Tensor):
 
         self._labels = labels   # assign the label
 
+    # TODO remove try/ except thing IMPORTANT
+    # make the label None of default
     def clone(self, *args, **kwargs):
         """
         Clone the LabelTensor. For more details, see
@@ -96,7 +98,12 @@ class LabelTensor(torch.Tensor):
         :return: a copy of the tensor
         :rtype: LabelTensor
         """
-        return LabelTensor(super().clone(*args, **kwargs), self.labels)
+        try:
+            out = LabelTensor(super().clone(*args, **kwargs), self.labels)
+        except:
+            out = super().clone(*args, **kwargs)
+
+        return out
 
     def to(self, *args, **kwargs):
         """
