### 2023-06-12 f35ee61
union domain implementation
- main file
- docstrings
