+++ .github/workflows/testing_pr.yml
@@ -4,6 +4,7 @@ on:
   pull_request:
     branches:
       - "master"
+      - "v0.1"
 
 
 jobs:
@@ -13,7 +14,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [windows-latest, macos-latest, ubuntu-latest]
-        python-version: [3.7, 3.8]
+        python-version: [3.7, 3.8, 3.9, 3.10]
         
     steps:
     - uses: actions/checkout@v2
