### 2023-06-13 mathLab/PINA#102 @yorkiva
The regularizer in the PINN class only implements the L2 regularization. It should also be extended to incorporate L1 regularization. I have had some experience with comparing these regularizations and I have found that for some problems L1 regularization can be quite useful when the training data (i.e. boundary/initial condition) is noisy.
Since the authors mention that a general functionality for regularization would be added in the upcoming release, this issue can be closed for now.
