### 2023-06-13 mathLab/PINA#112 @yorkiva
Review of the State of the Field for the JOSS paper
In the paper outline, statement of need section, some of the existing libraries have been mentioned. But their comparison with PINA seems a little underrepresented.

**Describe the solution you'd like**
Some additional description should be included detailing what novelty PINA brings. Stating that "PINA wants to emerge for its easiness of usage, allowing the users to quickly formulate the problem at hand and solve it, resulting in an intuitive frameworks designed by researchers for researchers." is inadequate. To some extent, the following points should be addressed-
- Is the codebase more transparent or intuitive to make it generalizable?
- If it indeed is intended for researchers, an important focus should be the addition of novel features (architectures, loss functions, regularizations, numerically difficult-to-solve problems) in new research projects. How easy or hard is it to add novel features compared to some of the existing software?
- Reusability of trained models for future inference (I think the repo could also benefit from some examples/tutorials showing this)
- Sustainability of the project (how will it be maintained down the line? what new features we might be expecting soon?)

I understand addressing all these issues might be difficult and lengthy, so I won't be very rigorous about addressing all of these issues. But some improvement is expected.

**Additional context**
Related to https://github.com/openjournals/joss-reviews/issues/5352
