### 2023-06-13 2447b1a
scheduler for all torch versions
