### 2023-06-13 d9d626a
Triangular Domain
-added __init__
-added is_inside()
-still need tests
