### 2023-06-13 836df75
TRIANGULAR DOMAIN -added tests -added necessary lines to __init__.py files
