+++ pina/geometry/__init__.py
@@ -2,9 +2,11 @@ __all__ = [
     'Location',
     'CartesianDomain',
     'EllipsoidDomain',
+    'TriangularDomain',
 ]
 
 from .location import Location
 from .cartesian import CartesianDomain
 from .ellipsoid import EllipsoidDomain
+from .triangular import TriangularDomain
 from .difference_domain import Difference
