### 2023-06-13 0e4ce85
check train test (#111)
* Update setup.py
* scheduler for all torch versions
---------

Co-authored-by: Dario Coscia <dariocoscia@cli-10-110-0-208.WIFIeduroamSTUD.units.it>
Co-authored-by: Nicola Demo <demo.nicola@gmail.com>
