### 2023-06-13 17ea3a0
Union and Init edits - minor edits to union - added Union to __init__.py of geometry
