+++ pina/geometry/__init__.py
@@ -2,9 +2,11 @@ __all__ = [
     'Location',
     'CartesianDomain',
     'EllipsoidDomain',
+    'Union'
 ]
 
 from .location import Location
 from .cartesian import CartesianDomain
 from .ellipsoid import EllipsoidDomain
 from .difference_domain import Difference
+from .union import Union
