### 2023-06-13 a482600
Implemented Union Tests - CartesianDomains - EllipsoidDomain - CartesianDomain and EllipsoidDomain
