### 2023-06-13 8e6ae16
Renaming Union - renamed tests cases - add param type for __init__
