+++ tests/test_union.py
@@ -4,22 +4,22 @@ import pytest
 from pina import LabelTensor, Union, EllipsoidDomain, CartesianDomain
 
 
-def test_constructor1():
+def test_constructor_two_CartesianDomains():
     Union([CartesianDomain({'x': [0, 1], 'y': [0, 1]}),
            CartesianDomain({'x': [0.5, 2], 'y': [-1, 0.1]})])
 
 
-def test_constructor2():
+def test_constructor2_two_EllipsoidDomains():
     Union([EllipsoidDomain({'x': [-1, 1], 'y': [-1, 1], 'z': [-1, 1]}),
            EllipsoidDomain({'x': [-0.5, 0.5], 'y': [-0.5, 0.5], 'z': [-0.5, 0.5]})])
 
 
-def test_constructor3():
+def test_constructor_EllipsoidDomain_CartesianDomain():
     Union([EllipsoidDomain({'x': [-1, 1], 'y': [-1, 1]}),
            CartesianDomain({'x': [-0.5, 0.5], 'y': [-0.5, 0.5]})])
 
 
-def test_is_inside1():
+def test_is_inside_two_CartesianDomains():
     pt_1 = LabelTensor(torch.tensor([[0.5, 0.5]]), ['x', 'y'])
     pt_2 = LabelTensor(torch.tensor([[-1, -1]]), ['x', 'y'])
     domain = Union([CartesianDomain({'x': [0, 1], 'y': [0, 1]}),
@@ -28,7 +28,7 @@ def test_is_inside1():
     assert domain.is_inside(pt_2) == False
 
 
-def test_is_inside2():
+def test_is_inside_two_EllipsoidDomains():
     pt_1 = LabelTensor(torch.tensor([[0.5, 0.5, 0.5]]), ['x', 'y', 'z'])
     pt_2 = LabelTensor(torch.tensor([[-1, -1, -1]]), ['x', 'y', 'z'])
     domain = Union([EllipsoidDomain({'x': [-1, 1], 'y': [-1, 1], 'z': [-1, 1]}),
@@ -37,7 +37,7 @@ def test_is_inside2():
     assert domain.is_inside(pt_2) == False
 
 
-def test_is_inside3():
+def test_is_inside_EllipsoidDomain_CartesianDomain():
     pt_1 = LabelTensor(torch.tensor([[0.5, 0.5]]), ['x', 'y'])
     pt_2 = LabelTensor(torch.tensor([[-1, -1]]), ['x', 'y'])
     domain = Union([EllipsoidDomain({'x': [-1, 1], 'y': [-1, 1], }),
