### 2023-06-13 3e5831b
Revamped tutorial 1 -editted -CartesianDomain() -trainer -sampling
