+++ tutorials/tutorial1/tutorial.py
@@ -3,19 +3,19 @@
 
 # # Tutorial 1: Physics Informed Neural Networks on PINA
 
-# In this tutorial we will show the typical use case of PINA on a toy problem. Specifically, the tutorial aims to introduce the following topics:
+# In this tutorial, we will demonstrate a typical use case of PINA on a toy problem. Specifically, the tutorial aims to introduce the following topics:
 # 
 # * Defining a PINA Problem,
-# * Build a `pinn` object,
-# * Sample points in the domain.
+# * Building a `pinn` object,
+# * Sampling points in a domain
 # 
-# These are the three main steps needed **before** training a Physics Informed Neural Network (PINN). We will show in detailed each step, and at the end we will solve a very simple problem with PINA.
+# These are the three main steps needed ***before** training a Physics Informed Neural Network (PINN). We will show each step in detail, and at the end, we will solve the problem.
 
 # ## PINA Problem
 
-# ### Initialize the Problem class
+# ### Initialize the `Problem` class
 
-# The problem definition in the PINA framework is done by building a phython `class`, inherited from one or more problem classes (`SpatialProblem`, `TimeDependentProblem`, `ParametricProblem`), depending on the nature of the problem treated. Let's see an example to better understand:
+# Problem definition in the PINA framework is done by building a python `class`, which inherits from one or more problem classes (`SpatialProblem`, `TimeDependentProblem`, `ParametricProblem`) depending on the nature of the problem. Below is an example:
 # #### Simple Ordinary Differential Equation
 # Consider the following:
 # 
@@ -28,53 +28,59 @@
 # \end{equation}
 # $$
 # 
-# with analytical solution $u(x) = e^x$. In this case we have that our ODE depends only on the spatial variable $x\in(0,1)$ , this means that our problem class is going to be inherited from `SpatialProblem` class:
+# with the analytical solution $u(x) = e^x$. In this case, our ODE depends only on the spatial variable $x\in(0,1)$ , meaning that our `Problem` class is going to be inherited from the `SpatialProblem` class:
 # 
 # ```python
 # from pina.problem import SpatialProblem
-# from pina import Span
+# from pina import CartesianProblem
 # 
 # class SimpleODE(SpatialProblem):
 #     
 #     output_variables = ['u']
-#     spatial_domain = Span({'x': [0, 1]})
+#     spatial_domain = CartesianProblem({'x': [0, 1]})
 # 
 #     # other stuff ...
 # ```
 # 
-# Notice that we define `output_variables` as a list of symbols, indicating the output variables of our equation (in this case only $u$). The `spatial_domain` variable indicates where the sample points are going to be sampled in the domain, in this case $x\in(0,1)$.
+# Notice that we define `output_variables` as a list of symbols, indicating the output variables of our equation (in this case only $u$). The `spatial_domain` variable indicates where the sample points are going to be sampled in the domain, in this case $x\in[0,1]$.
 
-# What about if we also have a time depencency in the equation? Well in that case our `class` will inherit from both `SpatialProblem` and `TimeDependentProblem`:
-# ```python
-# from pina.problem import SpatialProblem, TimeDependentProblem
-# from pina import Span
-# 
-# class TimeSpaceODE(SpatialProblem, TimeDependentProblem):
-#     
-#     output_variables = ['u']
-#     spatial_domain = Span({'x': [0, 1]})
-#     temporal_domain = Span({'x': [0, 1]})
-# 
-#     # other stuff ...
-# ```
-# where we have included the `temporal_domain` variable indicating the time domain where we want the solution.
-# 
-# Summarizing, in PINA we can initialize a problem with a class which is inherited from three base classes: `SpatialProblem`, `TimeDependentProblem`, `ParametricProblem`, depending on the type of problem we are considering. For reference:
[…]
