+++ docs/source/_rst/tutorial1/tutorial.rst
@@ -1,30 +1,29 @@
 Tutorial 1: Physics Informed Neural Networks on PINA
 ====================================================
 
-In this tutorial we will show the typical use case of PINA on a toy
-problem. Specifically, the tutorial aims to introduce the following
+In this tutorial, we will demonstrate a typical use case of PINA on a
+toy problem. Specifically, the tutorial aims to introduce the following
 topics:
 
 -  Defining a PINA Problem,
--  Build a ``pinn`` object,
--  Sample points in the domain.
+-  Building a ``pinn`` object,
+-  Sampling points in a domain
 
-These are the three main steps needed **before** training a Physics
-Informed Neural Network (PINN). We will show in detailed each step, and
-at the end we will solve a very simple problem with PINA.
+These are the three main steps needed \*\ **before** training a Physics
+Informed Neural Network (PINN). We will show each step in detail, and at
+the end, we will solve the problem.
 
 PINA Problem
 ------------
 
-Initialize the Problem class
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+Initialize the ``Problem`` class
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-The problem definition in the PINA framework is done by building a
-phython ``class``, inherited from one or more problem classes
-(``SpatialProblem``, ``TimeDependentProblem``, ``ParametricProblem``),
-depending on the nature of the problem treated. Let’s see an example to
-better understand: #### Simple Ordinary Differential Equation Consider
-the following:
+Problem definition in the PINA framework is done by building a python
+``class``, which inherits from one or more problem classes
+(``SpatialProblem``, ``TimeDependentProblem``, ``ParametricProblem``)
+depending on the nature of the problem. Below is an example: #### Simple
+Ordinary Differential Equation Consider the following:
 
 .. math::
 
@@ -36,20 +35,20 @@ the following:
    \end{cases}
    \end{equation}
 
-with analytical solution :math:`u(x) = e^x`. In this case we have that
-our ODE depends only on the spatial variable :math:`x\in(0,1)` , this
-means that our problem class is going to be inherited from
+with the analytical solution :math:`u(x) = e^x`. In this case, our ODE
+depends only on the spatial variable :math:`x\in(0,1)` , meaning that
+our ``Problem`` class is going to be inherited from the
 ``SpatialProblem`` class:
 
 .. code:: python
 
    from pina.problem import SpatialProblem
-   from pina import Span
+   from pina import CartesianProblem
 
    class SimpleODE(SpatialProblem):
        
        output_variables = ['u']
-       spatial_domain = Span({'x': [0, 1]})
+       spatial_domain = CartesianProblem({'x': [0, 1]})
 
        # other stuff ...
 
@@ -57,50 +56,52 @@ Notice that we define ``output_variables`` as a list of symbols,
 indicating the output variables of our equation (in this case only
 :math:`u`). The ``spatial_domain`` variable indicates where the sample
 points are going to be sampled in the domain, in this case
-:math:`x\in(0,1)`.
+:math:`x\in[0,1]`.
 
-What about if we also have a time depencency in the equation? Well in
-that case our ``class`` will inherit from both ``SpatialProblem`` and
+What about if our equation is also time dependent? In this case, our
+``class`` will inherit from both ``SpatialProblem`` and
 ``TimeDependentProblem``:
 
-.. code:: python
+.. code:: ipython3
 
-   from pina.problem import SpatialProblem, TimeDependentProblem
-   from pina import Span
+    from pina.problem import SpatialProblem, TimeDependentProblem
+    from pina import CartesianDomain
+    
[…]
