### 2023-06-13 f3c082e
Added Checks - added a location check in __init__ - return a label tensor in sample with labels 'x0', 'x1'...
