### 2023-06-14 mathLab/PINA#112 @dario-coscia
👋🏻 @yorkiva. Thank you for the very detailed comment, we appreciate the feedback. We will address the points you identified and make the corrections in the article in the following weeks.
