### 2023-06-14 12a748f
making union more pythonic
