+++ tests/test_union.py
@@ -1,7 +1,7 @@
 import torch
-import pytest
 
-from pina import LabelTensor, Union, EllipsoidDomain, CartesianDomain
+from pina import LabelTensor
+from pina.geometry import Union, EllipsoidDomain, CartesianDomain
 
 
 def test_constructor_two_CartesianDomains():
