+++ pina/__init__.py
@@ -7,6 +7,7 @@ __all__ = [
     'Location',
     'CartesianDomain',
     'TriangularDomain',
+    'Equation',
 ]
 
 from .meta import *
@@ -15,6 +16,5 @@ from .pinn import PINN
 from .trainer import Trainer
 from .plotter import Plotter
 from .condition import Condition
-from .geometry import Location
-from .geometry import CartesianDomain
-from .geometry import TriangularDomain
+from .geometry import Location, CartesianDomain, TriangularDomain
+from .equation.equation import Equation
