+++ pina/geometry/__init__.py
@@ -9,4 +9,4 @@ from .location import Location
 from .cartesian import CartesianDomain
 from .ellipsoid import EllipsoidDomain
 from .difference_domain import Difference
-from .union import Union
+from .union_domain import Union
