### 2023-06-14 3dba63d
small fixes
- make `geometries` private
- formatting
