### 2023-06-14 eaeb5c7
union domain implementation (#109)
* Union Class
* Implemented Union Tests

---------

Co-authored-by: Dario Coscia <93731561+dario-coscia@users.noreply.github.com>
