### 2023-06-14 68c6fe3
fix doc in residual block
