### 2023-06-15 691406c
more pythonic
