### 2023-06-15 5c23bea
refactored utils
