### 2023-06-16 5c88814
added iter and removed redundancy
