### 2023-06-16 40ef15a
refactored utils
