### 2023-06-16 f904a1f
refactored utils to try except
