### 2023-06-19 b76412d
enhanced difference domain
