### 2023-06-19 fdd8fae
removed object_name from all files
