### 2023-06-19 739e49f
added tests for check_consistency
