+++ pina/utils.py
@@ -18,7 +18,7 @@ def check_consistency(object, object_instance, subclass=False):
        ``'subclass=True'`` we check if the object is subclass
        if the ``'object_instance'``.
 
-    :param (iterable or class object )Object object: The object to check the inheritance 
+    :param (iterable or class object) object: The object to check the inheritance 
     :param Object object_instance: The parent class from where the object
         is expected to inherit
     :param str object_name: The name of the object
