+++ tests/test_utils.py
@@ -3,15 +3,10 @@ import torch
 from pina.utils import merge_tensors
 from pina.label_tensor import LabelTensor
 from pina import LabelTensor
-from pina.geometry import Union, EllipsoidDomain, CartesianDomain
+from pina.geometry import EllipsoidDomain, CartesianDomain
 from pina.utils import check_consistency
 import pytest
 from pina.geometry import Location
-
-cart1 = CartesianDomain({'x': [0, 1], 'y': [0, 1]})
-ellipsoid1 = EllipsoidDomain({'x': [1, 2], 'y': [-2, 1]})
-union = Union([cart1, ellipsoid1])
-example_input_pts = LabelTensor(torch.tensor([[0, 0, 0]]), ['x', 'y', 'z'])
 
 
 def test_merge_tensors():
@@ -28,15 +23,21 @@ def test_merge_tensors():
 
 
 def test_check_consistency_correct():
+    ellipsoid1 = EllipsoidDomain({'x': [1, 2], 'y': [-2, 1]})
+    example_input_pts = LabelTensor(torch.tensor([[0, 0, 0]]), ['x', 'y', 'z'])
+    
     check_consistency(example_input_pts, torch.Tensor)
-    check_consistency(union, Location)
+    check_consistency(CartesianDomain, Location, subclass=True)
     check_consistency(ellipsoid1, Location)
 
 
 def test_check_consistency_incorrect():
+    ellipsoid1 = EllipsoidDomain({'x': [1, 2], 'y': [-2, 1]})
+    example_input_pts = LabelTensor(torch.tensor([[0, 0, 0]]), ['x', 'y', 'z'])
+    
     with pytest.raises(ValueError):
         check_consistency(example_input_pts, Location)
     with pytest.raises(ValueError):
-        check_consistency(union, torch.Tensor)
+        check_consistency(ellipsoid1, Location, subclass=True)
     with pytest.raises(ValueError):
         check_consistency(ellipsoid1, torch.Tensor)
