+++ pina/geometry/difference_domain.py
@@ -6,11 +6,22 @@ from ..label_tensor import LabelTensor
 
 
 class Difference(Location):
-    """
-    """
+    """ PINA implementation of Difference of Domains."""
 
     def __init__(self, geometries):
+        """ PINA implementation of Difference of Domains.
 
+        :param list geometries: A list of geometries from 'pina.geometry' 
+            such as 'EllipsoidDomain' or 'CartesianDomain'.
+
+        :Example:
+            # Create two ellipsoid domains
+            >>> ellipsoid1 = EllipsoidDomain({'x': [-1, 1], 'y': [-1, 1]})
+            >>> ellipsoid2 = EllipsoidDomain({'x': [0, 2], 'y': [0, 2]})
+
+            # Create a difference of the ellipsoid domains
+            >>> difference = Difference([ellipsoid1, ellipsoid2])
+        """
         super().__init__()
 
         # difference checks
@@ -49,7 +60,7 @@ class Difference(Location):
         :param point: Point to be checked.
         :type point: torch.Tensor   
         :param bool check_border: If True, the border is considered inside.
-        :return: True if the point is inside the union domain, False otherwise.
+        :return: True if the point is inside the difference domain, False otherwise.
         :rtype: bool
         """
         flag = 0
@@ -59,35 +70,117 @@ class Difference(Location):
         return flag == 1
 
     def sample(self, n, mode='random', variables='all'):
-        sampled_points = []
+        """Sample routine.
+
+        :param n: Number of points to sample in the shape.
+        :type n: int
+        :param mode: Mode for sampling, defaults to 'random'.
+            Available modes include: random sampling, 'random'.
+        :type mode: str, optional
+        :param variables: pinn variable to be sampled, defaults to 'all'.
+        :type variables: str or list[str], optional
+
+        :Example:
+            # Create two Cartesian domains
+            >>> cartesian1 = CartesianDomain({'x': [0, 2], 'y': [0, 2]})
+            >>> cartesian2 = CartesianDomain({'x': [1, 3], 'y': [1, 3]})
+
+            # Create a difference of the ellipsoid domains
+            >>> difference = Difference([cartesian1, cartesian2])
+
+            >>> difference.sample(n=1000)
+                LabelTensor([[1.6802, 0.7723],
+                    [1.8872, 0.9972],
+                    [0.3166, 0.1213],
+                    ...,
+                    [2.1999, 2.0767],
+                    [1.0655, 2.7711],
+                    [1.6751, 2.3695]])
+
+            >>> len(difference.sample(n=1000)
+                1000
+
+        """
+        sampled = []
+
+        # calculate the number of points to sample for each geometry and the remainder
         remainder = n % len(self.geometries)
         num_points = n // len(self.geometries)
 
+        # sample the points
         for i, geometry in enumerate(self.geometries):
             if i < remainder:
                 num_points += 1
-            points = geometry.sample(num_points, mode, variables)
+            sampled_points = []
+            # makes sure point is uniquely inside 1 shape
+            while len(sampled_points) < num_points:
+                sample = geometry.sample(1, 'random')
+                # if not self.is_inside(sample) --> will be the intersection
+                if self.is_inside(sample):
+                    sampled_points.append(sample)
+            sampled.extend(sampled_points)
 
-            for point in points:
-                point.labels = [f'{i}' for i in self.variables]
-                if self.is_inside(point):
-                    sampled_points.append(point)
+        tensors = [point.data for point in sampled]
+        concatenated_tensor = torch.cat(tensors, dim=0)
 
-        return LabelTensor(torch.cat(sampled_points), labels=[f'{i}' for i in self.variables])
[…]
