+++ pina/geometry/difference_domain.py
@@ -25,7 +25,7 @@ class Difference(Location):
         super().__init__()
 
         # difference checks
-        check_consistency(geometries, Location)
+        check_consistency(geometries, Location, "test")
         self._check_difference_dimesions(geometries)
 
         # assign geometries
