+++ tutorials/tutorial 5/tutorial2.py
@@ -1,0 +1,161 @@
+import torch
+import torch.nn as nn
+
+from pina.problem import SpatialProblem
+from pina.operators import nabla, grad, div
+from pina.geometry import CartesianDomain
+from pina import Condition, LabelTensor, PINN
+from pina.trainer import Trainer
+from pina.model import FeedForward
+from pina.equation.equation import Equation
+from pina.equation.system_equation import SystemEquation
+from pina.equation.equation_factory import FixedValue
+from pina.plotter import Plotter
+from lightning.pytorch import seed_everything
+seed_everything(42, workers=True)
+
+# Define material
+E = 7
+nu = 0.3
+p = 'plain_strain'
+
+if p == 'plain_strain':  ### plain strain
+    lmbda = E * nu / (1 + nu) / (1 - 2 * nu)
+    mu = E / (1 + nu) / 2
+elif p == 'plain_stress':  ### plain stress
+    lmbda = E * nu / (1 + nu) / (1 - nu)
+    mu = E / (1 + nu) / 2
+
+def material(input_, output_):
+    u_grad = grad(output_, input_)
+
+    u1_xx = grad(u_grad, input_, components=['du1dx'], d=['x'])
+    u1_xy = grad(u_grad, input_, components=['du1dx'], d=['y'])
+    u1_yx = grad(u_grad, input_, components=['du1dy'], d=['x'])
+    u1_yy = grad(u_grad, input_, components=['du1dy'], d=['y'])
+
+    u2_xx = grad(u_grad, input_, components=['du2dx'], d=['x'])
+    u2_xy = grad(u_grad, input_, components=['du2dx'], d=['y'])
+    u2_yx = grad(u_grad, input_, components=['du2dy'], d=['x'])
+    u2_yy = grad(u_grad, input_, components=['du2dy'], d=['y'])
+
+    ### Calculate strain
+    e11 = u_grad.extract(['du1dx'])
+    e22 = u_grad.extract(['du2dy'])
+    e12 = 0.5 * (u_grad.extract(['du1dy']) + u_grad.extract(['du2dx']))
+
+    ### Calculate stress
+    s11 = (2 * mu + lmbda) * e11 + lmbda * e22
+    s22 = (2 * mu + lmbda) * e22 + lmbda * e11
+    s12 = 2 * mu * e12
+
+    ### Calculate equilibrium
+    Gex = (2 * mu + lmbda) * u1_xx + mu * u1_yy + mu * u2_xy + lmbda * u2_yx
+    Gey = (2 * mu + lmbda) * u2_yy + mu * u2_xx + mu*  u1_yx + lmbda * u1_xy
+
+    #sxx_x = lmbda * (u1_xx + u2_yx) + 2 * mu * u1_xx
+    #sxy_y = mu * (u1_yy + u2_xy)
+    #syy_y = lmbda * (u1_xy + u2_yy) + 2 * mu * u2_yy
+    #syx_x = mu * (u2_xx + u1_yx)
+    return e11, e22, e12, s11, s22, s12, Gex, Gey
+
+def equilibrium(input_, output_):
+    _, _, _, _, _, _, Gex, Gey = material(input_, output_)
+    return torch.stack([Gex, Gey], dim=1).squeeze()
+
+def gamma_left(input_, output_):
+    _, _, _, _, s11, s12, _, _ = material(input_, output_)
+    return torch.stack([s11, s12], dim=1).squeeze()
+
+def gamma_right(input_, output_):
+    _, _, _, s11, _, s12, _, _ = material(input_, output_)
+    return   torch.stack([s11 - torch.cos(torch.pi * input_.extract(['y'])/2), s12], dim=1).squeeze()
+
[…]
