### 2023-06-20 62a5e78
TriangleDomain -is_inside() -sample() -tests
