+++ pina/geometry/triangular.py
@@ -1,10 +1,8 @@
 import torch
 import numpy as np
-import math
 
+from pina.geometry import CartesianDomain
 from .location import Location
-from ..label_tensor import LabelTensor
-from ..utils import torch_lhs, chebyshev_roots
 
 
 class TriangularDomain(Location):
@@ -21,12 +19,32 @@ class TriangularDomain(Location):
             >>> spatial_domain = TriangularDomain({'vertex1': [0, 0], 'vertex2': [1, 1], 'vertex3': [0, 2]})
         """
 
+        if len(span_dict) <= 2:
+            raise ValueError("Too few vertices for a triangular domain")
+
+        if len(span_dict) == 3:  # 2D
+            self.dimension = 2
+
+            if not all(len(vertex) == 2 for vertex in span_dict.values()):
+                raise ValueError("Unsupported dimensions")
+
+        if len(span_dict) == 4:  # 3D
+            self.dimension = 3
+
+            if not all(len(vertex) == 3 for vertex in span_dict.values()):
+                raise ValueError("Unsupported dimensions")
+
+        if len(span_dict) >= 5:  # Won't handle greater than 3 dimensions
+            raise ValueError("Too many vertices for a triangular domain")
+
         self.fixed_, self.range_ = {}, {}
 
         for k, v in span_dict.items():
             if isinstance(v, (int, float)):
                 self.fixed_[k] = v
-            elif isinstance(v, (list, tuple)) and len(v) >= 2: # length might be able to be > 2
+            elif (
+                isinstance(v, (list, tuple)) and len(v) >= 2
+            ):  # length might be able to be > 2
                 self.range_[k] = v
             else:
                 raise TypeError
@@ -41,18 +59,18 @@ class TriangularDomain(Location):
         """
 
         return list(self.fixed_.keys()) + list(self.range_.keys())
-    
+
     @property
     def vertices(self):
         """
         Vertices of triangle.
 
         :return: Vectors defined in '__init__()'
-        :rtype: list[list]
+        :rtype: tuple[list]
         """
 
-        return list(self.fixed_.values()) + list(self.range_.values())
-    
+        return tuple(self.fixed_.values()) + tuple(self.range_.values())
+
     @property
     def vectors(self):
         """
@@ -62,10 +80,11 @@ class TriangularDomain(Location):
         :rtype: list[tuple]
         """
 
-        vertices = self.vertices
-            
-        return [[vertices[i+1][j]-vertices[i][j] for j in range(len(vertices[0]))] for i in range(len(vertices) - 1)] + [[vertices[-1][j]-vertices[0][j] for j in range(len(vertices[0]))]]
-    
+        return [
+            [self.vertices[i][j] - self.vertices[0][j] for j in range(self.dimension)]
+            for i in range(1, self.dimension + 1)
+        ]
+
     def is_inside(self, point, check_border=False):
         """
         Check if a point is inside the triangle.
@@ -79,42 +98,129 @@ class TriangularDomain(Location):
         :rtype: bool
         """
 
-        def _gramian(vector1, vector2):
+        def _area(vertices):
             """
-            Computes Gramian from vectors for sake of finding area.
+            Given vertices of triangle, calculates area
 
-            :param vector1: a vector
-            :type vector1: a list
-            :param vector2: a vector
-            :type vector2: a list
-            :return: Returns the Gramian of two vectors
-            :rtype: np matrix
-            """
-
-            x = [vector1, vector2]
-
-            return np.matmul(x, np.transpose(x))
-        
-        def _afg(gramian):
-            """
-            Area from Gramian.
-            
-            :param gramian: a Gramian matrix
-            :type gramian: np matrix
-            :return: Returns the area derived from Gramian
+            :param vertices: Vertices of triangle
[…]
