### 2023-06-20 f77f8c1
Updating difference domain - update sample/ is_inside - create tests - tried to resolve merge conflicts
