### 2023-06-20 a1f0617
Triangle geometry * creating TriangleDomain * tests
