+++ pina/geometry/__init__.py
@@ -10,3 +10,4 @@ from .cartesian import CartesianDomain
 from .ellipsoid import EllipsoidDomain
 from .difference_domain import Difference
 from .union_domain import Union
+from .triangle import TriangleDomain
