### 2023-06-20 1fbf4ad
Triangle geometry * creating TriangleDomain * tests
