+++ pina/geometry/__init__.py
@@ -21,6 +21,10 @@ from .ellipsoid import EllipsoidDomain
 from .exclusion_domain import Exclusion
 from .intersection_domain import Intersection
 from .union_domain import Union
+<<<<<<< HEAD
 from .difference_domain import Difference
 from .operation_interface import OperationInterface
 from .simplex import SimplexDomain
+=======
+from .triangle import TriangleDomain
+>>>>>>> 053d570 (Triangle geometry)
