### 2023-06-20 48583d7
Triangle geometry * creating TriangleDomain * tests
