+++ pina/geometry/triangle.py
@@ -1,0 +1,240 @@
+import torch
+from .location import Location
+from pina.geometry import CartesianDomain
+from pina import LabelTensor
+from ..utils import check_consistency
+
+class SimplexDomain(Location):
+    """PINA implementation of a Simplex."""
+
+    def __init__(self, simplex_matrix, sample_surface=False):
+        """
+        :param simplex_matrix: A matrix of LabelTensor objects representing
+            a vertex of the simplex (a tensor), and the coordinates of the
+            point (a list of labels).
+        :type simplex_matrix: list[LabelTensor]
+        :param sample_surface: A variable for choosing sample strategies. If
+            `sample_surface=True` only samples on the Simplex surface
+            frontier are taken. If `sample_surface=False`, no such criteria
+            is followed.
+        :type sample_surface: bool
+
+        .. warning::
+            Sampling for dimensions greater or equal to 10 could result
+            in a shrinking of the simplex, which degrades the quality
+            of the samples. For dimensions higher than 10, other algorithms
+            for sampling should be used.
+
+        :Example:
+
+            >>> spatial_domain = SimplexDomain(
+                    [
+                        LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+                        LabelTensor(torch.tensor([[1, 1]]), labels=["x", "y"]),
+                        LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+                    ], sample_surface = True
+                )
+        """
+
+        # check consistency of sample_surface
+        check_consistency(sample_surface, bool)
+        self._sample_surface = sample_surface
+
+        # check consistency of simplex_dict
+        if not isinstance(simplex_matrix, list):
+            raise ValueError(f"{type(simplex_matrix).__name__} must be {list}.")
+        check_consistency(simplex_matrix, LabelTensor)
+
+        # check consistency of labels
+        if simplex_matrix[0].labels != simplex_matrix[1].labels:
+            raise ValueError(f"Labels don't match.")
+        
+        # vertices, vectors, coordinates
+        self._vertices_matrix = torch.tensor(
+            [
+                [float(vertex.extract(label)) for label in vertex.labels]
+                for vertex in simplex_matrix
+            ]
+        ).T
+        self._vectors_shifted = (
+            self._vertices_matrix[:, :-1] - self._vertices_matrix[:, None, -1]
+        )
+        self._coordinates = simplex_matrix[0].labels
+
+        # build cartesian_bound
+        self._cartesian_bound = self._build_cartesian(self.variables)
+
+    @property
+    def coordinates(self):
+        return self._coordinates
+
+    @property
+    def variables(self):
+        return self._vertices_matrix
+
+    def _build_cartesian(self, vertices):
+        """
+        Build Cartesian border for Simplex domain to be used in sampling.
+
+        :param vertex_matrix: matrix of vertices
+        :type vertices: list[list]
+        :return: Cartesian border for triangular domain
+        :rtype: CartesianDomain
+        """
+
+        span_dict = {}
+
+        for i, coord in enumerate(self._coordinates):
+            sorted_vertices = sorted(vertices, key=lambda vertex: vertex[i])
+
+            # respective coord bounded by the lowest and highest values
+            span_dict[coord] = [sorted_vertices[0][i], sorted_vertices[-1][i]]
+
+        return CartesianDomain(span_dict)
+
+    def is_inside(self, point, check_border=False):
+        """
+        Check if a point is inside the simplex.
+        Uses the algorithm described involving barycentric coordinates:
+        https://en.wikipedia.org/wiki/Barycentric_coordinate_system
+
+        .. note::
+            When ```'sample_surface'``` in the ```'__init()__'```
+            is set to ```'True'```, then the method only checks
[…]
