+++ tutorials/tutorial1/tutorial.py
@@ -1,223 +1,0 @@
-#!/usr/bin/env python
-# coding: utf-8
-
-# # Tutorial 1: Physics Informed Neural Networks on PINA
-
-# In this tutorial we will show the typical use case of PINA on a toy problem. Specifically, the tutorial aims to introduce the following topics:
-# 
-# * Defining a PINA Problem,
-# * Build a `pinn` object,
-# * Sample points in the domain.
-# 
-# These are the three main steps needed **before** training a Physics Informed Neural Network (PINN). We will show in detailed each step, and at the end we will solve a very simple problem with PINA.
-
-# ## PINA Problem
-
-# ### Initialize the Problem class
-
-# The problem definition in the PINA framework is done by building a phython `class`, inherited from one or more problem classes (`SpatialProblem`, `TimeDependentProblem`, `ParametricProblem`), depending on the nature of the problem treated. Let's see an example to better understand:
-# #### Simple Ordinary Differential Equation
-# Consider the following:
-# 
-# $$
-# \begin{equation}
-# \begin{cases}
-# \frac{d}{dx}u(x) &=  u(x) \quad x\in(0,1)\\
-# u(x=0) &= 1 \\
-# \end{cases}
-# \end{equation}
-# $$
-# 
-# with analytical solution $u(x) = e^x$. In this case we have that our ODE depends only on the spatial variable $x\in(0,1)$ , this means that our problem class is going to be inherited from `SpatialProblem` class:
-# 
-# ```python
-# from pina.problem import SpatialProblem
-# from pina import Span
-# 
-# class SimpleODE(SpatialProblem):
-#     
-#     output_variables = ['u']
-#     spatial_domain = Span({'x': [0, 1]})
-# 
-#     # other stuff ...
-# ```
-# 
-# Notice that we define `output_variables` as a list of symbols, indicating the output variables of our equation (in this case only $u$). The `spatial_domain` variable indicates where the sample points are going to be sampled in the domain, in this case $x\in(0,1)$.
-
-# What about if we also have a time depencency in the equation? Well in that case our `class` will inherit from both `SpatialProblem` and `TimeDependentProblem`:
-# ```python
-# from pina.problem import SpatialProblem, TimeDependentProblem
-# from pina import Span
-# 
-# class TimeSpaceODE(SpatialProblem, TimeDependentProblem):
-#     
-#     output_variables = ['u']
-#     spatial_domain = Span({'x': [0, 1]})
-#     temporal_domain = Span({'x': [0, 1]})
-# 
-#     # other stuff ...
-# ```
-# where we have included the `temporal_domain` variable indicating the time domain where we want the solution.
-# 
-# Summarizing, in PINA we can initialize a problem with a class which is inherited from three base classes: `SpatialProblem`, `TimeDependentProblem`, `ParametricProblem`, depending on the type of problem we are considering. For reference:
-# * `SpatialProblem` $\rightarrow$ spatial variable(s) presented in the differential equation
-# * `TimeDependentProblem` $\rightarrow$ time variable(s) presented in the differential equation
-# * `ParametricProblem` $\rightarrow$ parameter(s) presented in the differential equation
-# 
-
-# ### Write the problem class
-# 
-# Once the problem class is initialized we need to write the differential equation in PINA language. For doing this we need to load the pina operators found in `pina.operators` module. Let's again consider the Equation (1) and try to write the PINA model class:
-
-# In[14]:
-
-
-from pina.problem import SpatialProblem
-from pina.operators import grad
-from pina import Condition, Span
-
-import torch
-
-
-class SimpleODE(SpatialProblem):
-
-    output_variables = ['u']
-    spatial_domain = Span({'x': [0, 1]})
-
-    # defining the ode equation
-    def ode_equation(input_, output_):
-
-        # computing the derivative
[…]
