### 2023-06-20 d225315
Difference Fixes - random iteration list for sample - added more commets - fixed num_points for loop
