+++ pina/geometry/difference_domain.py
@@ -16,13 +16,13 @@ class Difference(Location):
     def sample(self, n, mode='random', variables='all'):
         """
         """
-        assert mode is 'random', 'Only random mode is implemented'
+        assert mode == 'random', 'Only random mode is implemented'
 
         samples = []
         while len(samples) < n:
             sample = self.first.sample(1, 'random')
             if not self.second.is_inside(sample):
-                samples.append(sample.tolist()[0])
+                samples.append(sample)
 
         import torch
-        return LabelTensor(torch.tensor(samples), labels=['x', 'y'])
+        return LabelTensor(torch.cat(samples), labels=['x', 'y'])
