+++ pina/geometry/difference_domain.py
@@ -134,16 +134,6 @@ class Difference(Location):
         :param geometries: Geometries to be checked.
         :type geometries: list[Location]
         """
-        assert mode == 'random', 'Only random mode is implemented'
-
-        samples = []
-        while len(samples) < n:
-            sample = self.first.sample(1, 'random')
-            if not self.second.is_inside(sample):
-                samples.append(sample)
-
-        import torch
-        return LabelTensor(torch.cat(samples), labels=['x', 'y'])
         for geometry in geometries:
             if geometry.variables != geometries[0].variables:
                 raise NotImplementedError(
