### 2023-06-21 2f3a48b
TriangleDomain * converted numpy operations into torch * tests now all pass
