### 2023-06-21 5a0d586
TriangleDomain * converted numpy operations into torch * tests now all pass
