### 2023-06-21 5a75074
Minor Fixes - sped up sample in Difference - pass in necessary variables in sample and Location - added a sample size and 3 geometries test case
