+++ pina/geometry/difference_domain.py
@@ -124,7 +124,7 @@ class Difference(Location):
                 # if not self.is_inside(sample) --> will be the intersection
                 if self.is_inside(sample):
                     sampled_points.append(sample)
-                sampled += sampled_points
+            sampled += sampled_points
 
         return LabelTensor(torch.cat(sampled), labels=[f'{i}' for i in self.variables])
 
