### 2023-06-21 f12cbf9
added sampling check
