+++ pina/geometry/triangle.py
@@ -1,0 +1,251 @@
+import torch
+
+from pina.geometry import CartesianDomain
+from .location import Location
+from pina import LabelTensor
+from ..utils import check_consistency
+
+
+class TriangleDomain(Location):
+    """PINA implementation of a Triangle."""
+
+    def __init__(self, span_dict, labels, sample_surface=False):
+        """
+        :param span_dict: A dictionary with dict-key a string representing
+            the input variables for the pinn, and dict-value a list
+            representing vertices of triangle.
+        :type span_dict: dict
+
+        :Example:
+            >>> spatial_domain = TriangleDomain({'vertex1': [0, 0], 'vertex2': [1, 1], 'vertex3': [0, 2]})
+        """
+
+        # check consistency labels
+        check_consistency(labels, list)
+        for lab in labels:
+            check_consistency(lab, str)
+
+        self._labels = lables
+
+        # check consistency sample_surface
+        check_consistency(sample_surface, bool)
+        self._sample_surface = lables
+
+
+        # checks
+        # 1. check the span, also number entries = numb labels
+
+        # TODO lets condense
+        if len(span_dict) <= 2:
+            raise ValueError("Too few vertices for a triangular domain")
+
+        if len(span_dict) == 3:  # 2D
+            self.dimension = 2
+
+            if not all(len(vertex) == 2 for vertex in span_dict.values()):
+                raise ValueError("Unsupported dimensions")
+
+        if len(span_dict) == 4:  # 3D
+            self.dimension = 3
+
+            if not all(len(vertex) == 3 for vertex in span_dict.values()):
+                raise ValueError("Unsupported dimensions")
+
+        if len(span_dict) >= 5:  # Won't handle greater than 3 dimensions
+            raise ValueError("Too many vertices for a triangular domain")
+
+        self.fixed_, self.range_ = {}, {}
+
+        for k, v in span_dict.items():
+            if isinstance(v, (int, float)):
+                self.fixed_[k] = v
+            elif (
+                isinstance(v, (list, tuple)) and len(v) >= 2
+            ):  # length might be able to be > 2
+                self.range_[k] = v
+            else:
+                raise TypeError
+
+    @property
+    def check_border(self):
+        return self._check_border
+
+    @property
+    def variables(self):
+        """
+        TODO
+        """
+        return self._labels
+
+    @property
+    def vertices(self):
+        """
+        Vertices of triangle.
+
+        :return: Vectors defined in '__init__()'
+        :rtype: tuple[list]
+        """
+
+        return list(self.fixed_.values()) + list(self.range_.values())
+
+    @property
+    def vectors(self):
+        """
+        Vertices of triangular domain.
+
+        :return: Vertices
+        :rtype: list[tuple]
+        """
+        # have tensors in a matrix form numb_dim x numb_dim - 1 
+        return [
+            [self.vertices[i][j] - self.vertices[0][j] for j in range(self.dimension)]
+            for i in range(1, self.dimension + 1)
+        ]
+
+    def is_inside(self, point, check_border=False):
+        """
+        Check if a point is inside the triangle.
+
+        :param point: Point to be checked
+        :type point: LabelTensor
+        :param check_border: Check if the point is also on the frontier
+            of the triangle, default False.
+        :type check_border: bool
+        :return: Returning True if the point is inside, False otherwise.
+        :rtype: bool
+        """
+
+        def _area(vertices):
+            """
+            Given vertices of triangle, calculates area
+
+            :param vertices: Vertices of triangle
+            :type vertices: tuple
+            :return Returns area of triangle
+            :rtype: float
+            """
[…]
