+++ pina/geometry/triangle.py
@@ -3,12 +3,13 @@ import torch
 from pina.geometry import CartesianDomain
 from .location import Location
 from pina import LabelTensor
+from ..utils import check_consistency
 
 
 class TriangleDomain(Location):
     """PINA implementation of a Triangle."""
 
-    def __init__(self, span_dict):
+    def __init__(self, span_dict, labels, sample_surface=False):
         """
         :param span_dict: A dictionary with dict-key a string representing
             the input variables for the pinn, and dict-value a list
@@ -16,9 +17,25 @@ class TriangleDomain(Location):
         :type span_dict: dict
 
         :Example:
-            >>> spatial_domain = TriangularDomain({'vertex1': [0, 0], 'vertex2': [1, 1], 'vertex3': [0, 2]})
+            >>> spatial_domain = TriangleDomain({'vertex1': [0, 0], 'vertex2': [1, 1], 'vertex3': [0, 2]})
         """
 
+        # check consistency labels
+        check_consistency(labels, list)
+        for lab in labels:
+            check_consistency(lab, str)
+
+        self._labels = lables
+
+        # check consistency sample_surface
+        check_consistency(sample_surface, bool)
+        self._sample_surface = lables
+
+
+        # checks
+        # 1. check the span, also number entries = numb labels
+
+        # TODO lets condense
         if len(span_dict) <= 2:
             raise ValueError("Too few vertices for a triangular domain")
 
@@ -50,15 +67,15 @@ class TriangleDomain(Location):
                 raise TypeError
 
     @property
+    def check_border(self):
+        return self._check_border
+
+    @property
     def variables(self):
         """
-        Spatial variables.
-
-        :return: Spatial variables defined in '__init__()'
-        :rtype: list[str]
+        TODO
         """
-
-        return list(self.fixed_.keys()) + list(self.range_.keys())
+        return self._labels
 
     @property
     def vertices(self):
@@ -69,7 +86,7 @@ class TriangleDomain(Location):
         :rtype: tuple[list]
         """
 
-        return tuple(self.fixed_.values()) + tuple(self.range_.values())
+        return list(self.fixed_.values()) + list(self.range_.values())
 
     @property
     def vectors(self):
@@ -79,7 +96,7 @@ class TriangleDomain(Location):
         :return: Vertices
         :rtype: list[tuple]
         """
-
+        # have tensors in a matrix form numb_dim x numb_dim - 1 
         return [
             [self.vertices[i][j] - self.vertices[0][j] for j in range(self.dimension)]
             for i in range(1, self.dimension + 1)
@@ -195,25 +212,25 @@ class TriangleDomain(Location):
         # for 2D
 
         # Construct CartesianDomain that contains triangle
-        vertices_by_x = sorted(self.vertices, lambda vertex: vertex[0])
-        vertices_by_y = sorted(self.vertices, lambda vertex: vertex[1])
+        vertices_by_x = sorted(self.vertices, key = lambda vertex: vertex[0])
+        vertices_by_y = sorted(self.vertices, key = lambda vertex: vertex[1])
 
         if self.dimension == 3:
-            vertices_by_z = sorted(self.vertices, lambda vertex: vertex[2])
+            vertices_by_z = sorted(self.vertices, key = lambda vertex: vertex[2])
 
             circumscribing_domain = CartesianDomain(
                 {
-                    "x": [vertices_by_x[0], vertices_by_x[-1]],
-                    "y": [vertices_by_y[0], vertices_by_y[-1]],
-                    "z": [vertices_by_z[0], vertices_by_z[-1]],
+                    "x": [vertices_by_x[0][0], vertices_by_x[-1][0]],
[…]
