### 2023-06-21 4becbdb
Docs fix cartesian.py
