### 2023-06-21 18748dc
device for sample points in absProblem
