### 2023-06-21 e95327f
test for gpu training
