### 2023-06-21 44e2f15
Created Intersection - inherits from Exclusion - changed sample so it's the intersection - added Exclusion and Intersection to __init__
