### 2023-06-22 3c01341
New Geometries - created a Difference domain - cleaned up Intersection to aviod repetitive code - added Difference to __init__
