+++ pina/geometry/__init__.py
@@ -5,6 +5,7 @@ __all__ = [
     'Union',
     'Intersection',
     'Exclusion',
+    'Difference',
 ]
 
 from .location import Location
@@ -13,3 +14,4 @@ from .ellipsoid import EllipsoidDomain
 from .exclusion_domain import Exclusion
 from .intersection_domain import Intersection
 from .union_domain import Union
+from .difference_domain import Difference
