+++ pina/geometry/difference_domain.py
@@ -7,11 +7,11 @@ from ..label_tensor import LabelTensor
 import random
 
 
-class Intersection(Exclusion):
-    """ PINA implementation of Intersection of Domains."""
+class Difference(Exclusion):
+    """ PINA implementation of Difference of Domains."""
 
     def __init__(self, geometries):
-        """ PINA implementation of Intersection of Domains.
+        """ PINA implementation of Difference of Domains.
 
         :param list geometries: A list of geometries from 'pina.geometry' 
             such as 'EllipsoidDomain' or 'CartesianDomain'.
@@ -21,11 +21,15 @@ class Intersection(Exclusion):
             >>> ellipsoid1 = EllipsoidDomain({'x': [-1, 1], 'y': [-1, 1]})
             >>> ellipsoid2 = EllipsoidDomain({'x': [0, 2], 'y': [0, 2]})
 
-            # Create a Intersection of the ellipsoid domains
-            >>> intersection = Intersection([ellipsoid1, ellipsoid2])
+            # Create a Difference of the ellipsoid domains
+            >>> difference = Difference([ellipsoid1, ellipsoid2])
         """
         super().__init__(geometries)
 
+        # assign geometries
+        self.first = geometries[0]
+        self.rest = geometries[1:]
+    
     def sample(self, n, mode='random', variables='all'):
         """Sample routine.
 
@@ -42,10 +46,10 @@ class Intersection(Exclusion):
             >>> cartesian1 = CartesianDomain({'x': [0, 2], 'y': [0, 2]})
             >>> cartesian2 = CartesianDomain({'x': [1, 3], 'y': [1, 3]})
 
-            # Create a Intersection of the ellipsoid domains
-            >>> intersection = Intersection([cartesian1, cartesian2])
+            # Create a Difference of the ellipsoid domains
+            >>> difference = Difference([cartesian1, cartesian2])
 
-            >>> intersection.sample(n=1000)
+            >>> difference.sample(n=1000)
                 LabelTensor([[1.5562, 1.8656],
                             [1.1060, 1.2712],
                             [1.3909, 1.5579],
@@ -54,7 +58,7 @@ class Intersection(Exclusion):
                             [1.1835, 1.5107],
                             [1.1986, 1.6461]])
 
-            >>> len(Intersection.sample(n=1000)
+            >>> len(difference.sample(n=1000)
                 1000
 
         """
@@ -62,27 +66,21 @@ class Intersection(Exclusion):
             raise NotImplementedError(f'{mode} is not a valid mode for sampling.')
         
         sampled = []
-
-        # calculate the number of points to sample for each geometry and the remainder.
-        remainder = n % len(self.geometries)
-        num_points = n // len(self.geometries)
-
+        
         # sample the points
-        # NB. geometries as shuffled since if we sample
-        # multiple times just one point, we would end
-        # up sampling only from the first geometry.
-        iter_ = random.sample(self.geometries, len(self.geometries))
-        for i, geometry in enumerate(iter_):
-            sampled_points = []
-            # int(i < remainder) is one only if we have a remainder
-            # different than zero. Notice that len(geometries) is
-            # always smaller than remaider.
-            # makes sure point is uniquely inside 1 shape.
-            while len(sampled_points) < (num_points + int(i < remainder)):
-                sample = geometry.sample(1, mode, variables)
-                # if not self.is_inside(sample) --> will be the intersection
-                if not self.is_inside(sample):
-                    sampled_points.append(sample)
-            sampled += sampled_points
+        while len(sampled) < n:
+            # get sample point from first geometry
+            point = self.first.sample(1, mode, variables)
+            is_inside = False
+
+            # check if point is inside any other geometry
+            for geometry in self.rest:
[…]
