### 2023-06-22 4a57cc7
Test Cases - added Difference test cases - added Intersect test cases - prettied up code
