### 2023-06-22 507abe7
Solvers for multiple models - Implementing the possibility to add multiple models for solvers (e.g. GAN) - Implementing GAROM solver, see https://arxiv.org/abs/2305.15881 - Implementing tests for GAROM solver (cpu only) - Fixing docs PINNs - Creating a solver directory, for consistency in the package
