+++ pina/solvers/garom.py
@@ -1,0 +1,244 @@
+""" Module for PINN """
+import torch
+try:
+    from torch.optim.lr_scheduler import LRScheduler  # torch >= 2.0
+except ImportError:
+    from torch.optim.lr_scheduler import _LRScheduler as LRScheduler # torch < 2.0
+
+from torch.optim.lr_scheduler import ConstantLR
+from .solver import SolverInterface
+from ..utils import check_consistency
+from ..loss import LossInterface, LpLoss
+from torch.nn.modules.loss import _Loss
+
+
+class GAROM(SolverInterface):
+    """
+    GAROM solver class. This class implements Generative Adversarial
+    Reduced Order Model solver, using user specified ``models`` to solve
+    a specific order reduction``problem``. 
+
+    .. seealso::
+
+        **Original reference**: Coscia, D., Demo, N., & Rozza, G. (2023).
+        Generative Adversarial Reduced Order Modelling. 
+        arXiv preprint arXiv:2305.15881.
+        <https://doi.org/10.48550/arXiv.2305.15881>`_.
+    """
+    def __init__(self,
+                 problem,
+                 generator,
+                 discriminator,
+                 loss = LpLoss(p=1),
+                 optimizer_generator=torch.optim.Adam,
+                 optimizer_generator_kwargs={'lr' : 0.001},
+                 optimizer_discriminator=torch.optim.Adam,
+                 optimizer_discriminator_kwargs={'lr' : 0.001},
+                 scheduler_generator=ConstantLR,
+                 scheduler_generator_kwargs={"factor": 1, "total_iters": 0},
+                 scheduler_discriminator=ConstantLR,
+                 scheduler_discriminator_kwargs={"factor": 1, "total_iters": 0},
+                 gamma = 0.3,
+                 lambda_k = 0.001,
+                 regularizer = False
+                 ):
+        """
+        :param AbstractProblem problem: The formualation of the problem.
+        :param torch.nn.Module generator: The neural network model to use
+            for the generator.
+        :param torch.nn.Module discriminator: The neural network model to use
+            for the discriminator.
+        :param torch.nn.Module loss: The loss function used as minimizer,
+            default ``LpLoss(p=1)`` as in the original paper.
+        :param torch.optim.Optimizer optimizer_generator: The neural
+            network optimizer to use for the generator network
+            , default is `torch.optim.Adam`.
+        :param dict optimizer_generator_kwargs: Optimizer constructor keyword
+            args. for the generator.
+        :param torch.optim.Optimizer optimizer_discriminator: The neural
+            network optimizer to use for the discriminator network
+            , default is `torch.optim.Adam`.
+        :param dict optimizer_discriminator_kwargs: Optimizer constructor keyword
+            args. for the discriminator.
+        :param torch.optim.LRScheduler scheduler_generator: Learning
+            rate scheduler for the generator.
+        :param dict scheduler_generator_kwargs: LR scheduler constructor keyword args.
+        :param torch.optim.LRScheduler scheduler_discriminator: Learning
+            rate scheduler for the discriminator.
+        :param dict scheduler_discriminator_kwargs: LR scheduler constructor keyword args.
+        :param gamma: Ratio of expected loss for generator and discriminator, defaults to 0.3.
+        :type gamma: float, optional
+        :param lambda_k: Learning rate for control theory optimization, defaults to 0.001.
+        :type lambda_k: float, optional
+        :param regularizer: Regularization term in the GAROM loss, defaults to False.
+        :type regularizer: bool, optional
+
+        .. warning::
+            The algorithm works only for data-driven model. Hence in the ``problem`` definition
+            the codition must only contain ``input_points`` (e.g. coefficient parameters, time
[…]
