### 2023-06-22 25dbe82
docstrings and minor fixes
