### 2023-06-22 b405e7d
updated docstrings
