### 2023-06-22 89e0c49
intersection comments
