+++ pina/geometry/simplex.py
@@ -1,240 +1,127 @@
 import torch
-from .location import Location
-from pina.geometry import CartesianDomain
+import pytest
+
 from pina import LabelTensor
-from ..utils import check_consistency
+from pina.geometry import SimplexDomain
 
-class SimplexDomain(Location):
-    """PINA implementation of a Simplex."""
 
-    def __init__(self, simplex_matrix, sample_surface=False):
-        """
-        :param simplex_matrix: A matrix of LabelTensor objects representing
-            a vertex of the simplex (a tensor), and the coordinates of the
-            point (a list of labels).
-        :type simplex_matrix: list[LabelTensor]
-        :param sample_surface: A variable for choosing sample strategies. If
-            `sample_surface=True` only samples on the Simplex surface
-            frontier are taken. If `sample_surface=False`, no such criteria
-            is followed.
-        :type sample_surface: bool
-
-        .. warning::
-            Sampling for dimensions greater or equal to 10 could result
-            in a shrinking of the simplex, which degrades the quality
-            of the samples. For dimensions higher than 10, other algorithms
-            for sampling should be used.
-
-        :Example:
-
-            >>> spatial_domain = SimplexDomain(
-                    [
-                        LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
-                        LabelTensor(torch.tensor([[1, 1]]), labels=["x", "y"]),
-                        LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
-                    ], sample_surface = True
-                )
-        """
-
-        # check consistency of sample_surface
-        check_consistency(sample_surface, bool)
-        self._sample_surface = sample_surface
-
-        # check consistency of simplex_dict
-        if not isinstance(simplex_matrix, list):
-            raise ValueError(f"{type(simplex_matrix).__name__} must be {list}.")
-        check_consistency(simplex_matrix, LabelTensor)
-
-        # check consistency of labels
-        if simplex_matrix[0].labels != simplex_matrix[1].labels:
-            raise ValueError(f"Labels don't match.")
-        
-        # vertices, vectors, coordinates
-        self._vertices_matrix = torch.tensor(
+def test_constructor():
+    SimplexDomain(
+        [
+            LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[1, 1]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+        ]
+    )
+    SimplexDomain(
+        [
+            LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[1, 1]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+        ],
+        sample_surface=True,
+    )
+    with pytest.raises(ValueError):
+        SimplexDomain(
             [
-                [float(vertex.extract(label)) for label in vertex.labels]
-                for vertex in simplex_matrix
+                LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+                LabelTensor(torch.tensor([[1, 1]]), labels=["x", "z"]),
+                LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
             ]
-        ).T
-        self._vectors_shifted = (
[…]
