### 2023-06-23 df62a90
Committing edits after refactoring before Friday meeting * tests not fulling working * on_border has bugs   - only works for 2D   - many workarounds due to bugs in LabelTensor
