+++ pina/geometry/simplex.py
@@ -1,0 +1,321 @@
+import torch
+
+from pina.geometry import CartesianDomain
+from .location import Location
+from pina import LabelTensor
+from ..utils import check_consistency
+
+class SimplexDomain(Location):
+    """PINA implementation of a Simplex."""
+
+    def __init__(self, simplex_dict, labels, sample_surface=False):
+        """
+        :param simplex_dict: A dictionary with dict-key a string representing
+            the input variables for the pinn, and dict-value a list
+            representing vertices of the simplex.
+        :type simplex_dict: dict
+        :param labels: A list of labels for vertex components
+        :type labels: list[str]
+        :param sample_surface: A variable for choosing sample strategies. If
+            `sample_surface=True` only samples on the Simplex surface
+            frontier are taken. If `sample_surface=False`, no such criteria
+            is followed.
+        :type sample_surface: bool
+        
+        :Example:
+
+            >>> spatial_domain = SimplexDomain({'vertex1': [0, 0], 'vertex2': [1, 1], 'vertex3': [0, 2]}, ['x', 'y'])
+        """
+
+        # check consistency of labels
+        if not isinstance(labels, list):
+            raise ValueError(f"{type(labels).__name__} must be {list}.")
+        for lab in labels:
+            check_consistency(lab, str)
+        self._coordinate_labels = labels
+        
+        # check consistency of sample_surface
+        check_consistency(sample_surface, bool)
+        self._sample_surface = sample_surface
+
+        # check consistency of simplex_dict
+        check_consistency(simplex_dict, dict)
+        for vertex in simplex_dict.values():
+            if not isinstance(vertex, list):
+                raise ValueError(f"{type(vertex).__name__} must be {list}.")
+        
+
+        # vertices, vectors, dimension 
+        self._vertices = {label: LabelTensor(torch.tensor([vertex]), self.variables) for label, vertex in simplex_dict.items()}
+        self._vectors = self._basis_vectors(self._vertices_list)
+        self._dimension = len(simplex_dict)-1
+
+
+        # build cartesian_bound
+        self._cartesian_bound = self._build_cartesian(list(simplex_dict.values()), labels)
+
+
+    @property
+    def variables(self):
+        """
+        Coordinate labels of simplex.
+
+        :return: Coordinate labels
+        :rtype: list[str]
+        """
+
+        return self._coordinate_labels
+
+
+    @property
+    def vertices(self):
+        """
+        Vertices of simplex.
+
+        :return: Vectors
+        :rtype: list[list]
+        """
+
+        return self._vertices
+    
+
+    @property
+    def vectors(self):
+        """
+        Vectors.
+
+        :return: Vectors
+        :rtype: dict(LabelTensor)
+        """
+        return self._vectors
+    
+
+    @property
+    def _vertices_list(self):
+        """
+        List of vectors.
+
+        :return: List of vectors
+        :rtype: list[LabelTensor]
+        """
+        return list(self.vertices.values())
+    
+
+    @property
+    def cartesian_bound(self):
+        """
+        Cartesian border for Simplex domain.
+
+        :return: Cartesian border for Simplex domain
+        :rtype: CartesianDomain
+        """
+
+        return self._cartesian_bound
+    
+
+    @property
+    def dimension(self):
+        """
+        Dimension of Simplex domain.
+        
+        :return: dimension
+        :rtype: int
+        """
+
+        return self._dimension
+    
+
+    @property
+    def sample_surface(self):
+        """
+        Whether the surface should be sampled or not.
+
+        :return: Whether the surface should be sampled or not
+        :rtype: bool
+        """
+
+        return self._sample_surface
+    
+
+    def _basis_vectors(self, vertices):
+        """
+        Basis vectors for simplex.
+
+        :return: Basis vectors
[…]
