+++ pina/geometry/__init__.py
@@ -2,8 +2,8 @@ __all__ = [
     'Location',
     'CartesianDomain',
     'EllipsoidDomain',
-    'Union', 
-    'TriangleDomain'
+    'Union',
+    'SimplexDomain'
 ]
 
 from .location import Location
@@ -14,8 +14,4 @@ from .intersection_domain import Intersection
 from .union_domain import Union
 from .difference_domain import Difference
 from .union_domain import Union
-<<<<<<< HEAD
-from .triangle import SimplexDomain
-=======
-from .triangle import TriangleDomain
->>>>>>> 053d570 (Triangle geometry)
+from .simplex import SimplexDomain
