+++ pina/geometry/__init__.py
@@ -3,7 +3,7 @@ __all__ = [
     'CartesianDomain',
     'EllipsoidDomain',
     'Union', 
-    'TriangleDomain'
+    'SimplexDomain'
 ]
 
 from .location import Location
@@ -11,4 +11,4 @@ from .cartesian import CartesianDomain
 from .ellipsoid import EllipsoidDomain
 from .difference_domain import Difference
 from .union_domain import Union
-from .triangle import TriangleDomain
+from .simplex import SimplexDomain
