### 2023-06-23 443f67f
Update test_pinn.py
