+++ tests/test_pinn.py
@@ -106,7 +106,7 @@ def test_train_extra_feats_cpu():
     trainer = Trainer(solver=pinn, kwargs={'max_epochs' : 5, 'accelerator':'cpu'})
     trainer.train()
 
-def test_train_cpu():
+def test_train_gpu():
     poisson_problem = Poisson()
     boundaries = ['gamma1', 'gamma2', 'gamma3', 'gamma4']
     n = 10
