+++ tutorials/tutorial 5/tutorial3.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
 
-from pina.problem import SpatialProblem
+from pina.problem import SpatialProblem, TimeDependentProblem
 from pina.operators import grad
 from pina.geometry import CartesianDomain
 from pina import Condition, PINN
@@ -10,8 +10,6 @@ from pina.equation.equation import Equation
 from pina.equation.system_equation import SystemEquation
 from pina.plotter import Plotter
 from lightning.pytorch import seed_everything
-
-seed_everything(42, workers=True)
 
 # Define material
 E = 7
@@ -62,46 +60,48 @@ def equilibrium(input_, output_):
 
 
 def gamma_left(input_, output_):
-    _, _, _, _, s11, s12, _, _ = material(input_, output_)
-    return torch.stack([s11, s12], dim=1).squeeze()
+    _, _, _, s11, _, _, _, _ = material(input_, output_)
+    return s11
 
 
 def gamma_right(input_, output_):
-    _, _, _, s11, _, s12, _, _ = material(input_, output_)
-    return torch.stack([s11 -torch.cos(torch.pi * input_.extract(['y']) / 2), s12], dim=1).squeeze()
+    _, _, _, s11, _, _, _, _ = material(input_, output_)
+    return s11
 
 
 def gamma_bottom(input_, output_):
-    _, _, _, _, s22, s12, _, _ = material(input_, output_)
-    return torch.stack([s22, s12], dim=1).squeeze()
+    u1 = output_.extract(['u1'])
+    u2 = output_.extract(['u2'])
+    return torch.stack([u1, u2], dim=1).squeeze()
 
 
 def gamma_top(input_, output_):
-    _, _, _, _, s22, s12, _, _ = material(input_, output_)
-    return torch.stack([s22, s12], dim=1).squeeze()
+    u2 = output_.extract(['u2'])
+    return u2 - torch.sin(0.5 * torch.pi * input_.extract(['x']))*0.04
 
 
 class Mechanics(SpatialProblem):
     output_variables = ['u1', 'u2']
-    spatial_domain = CartesianDomain({'x': [0, 1], 'y': [0, 1]})
+    spatial_domain = CartesianDomain({'x': [-1, 1], 'y': [-1, 1]})
 
     conditions = {
         'gamma_left': Condition(
-            location=CartesianDomain({'x': 0, 'y': [0, 1]}),
+            location=CartesianDomain({'x': -1, 'y': [-1, 1]}),
             equation=Equation(gamma_left)),
         'gamma_right': Condition(
-            location=CartesianDomain({'x': 1, 'y': [0, 1]}),
+            location=CartesianDomain({'x': 1, 'y': [-1, 1]}),
             equation=Equation(gamma_right)),
-        'gamma_bottom': Condition(
-            location=CartesianDomain({'x': [0, 1], 'y': 0}),
-            equation=Equation(gamma_bottom)),
+        # 'gamma_bottom': Condition(
+        #    location=CartesianDomain({'x':[-0.5, 0.5], 'y': -0.5}),
+        #    equation=Equation(gamma_bottom)),
         'gamma_top': Condition(
-            location=CartesianDomain({'x': [0, 1], 'y': 1}),
+            location=CartesianDomain({'x': [-1, 1], 'y': 1}),
             equation=Equation(gamma_top)),
         'D': Condition(
-            location=CartesianDomain({'x': [0, 1], 'y': [0, 1]}),
+            location=CartesianDomain({'x': [-1, 1], 'y': [-1, 1]}),
             equation=SystemEquation([equilibrium]))
     }
+
 
 # make the problem
 bvp_problem = Mechanics()
[…]
