### 2023-06-26 9b1cd06
Fix ellipsoid.py bug
- Fixing a typo in `ellipsoid.py`
