### 2023-06-27 mathLab/PINA#140 @akshaysubr
Review of the State of the Field for the JOSS paper
In the statement of need section, some PyTorch and TensorFlow based libraries are mentioned. One notable omission in the PyTorch section is [Modulus](https://github.com/NVIDIA/modulus-sym). In addition the description of the unique value proposition of PINA is a little too high level.

**Describe the solution you'd like**
In addition to adding Modulus to the list of existing works, it would be good to add a couple of examples or description of specific features to support the claim: "PINA wants to emerge for its easiness of usage, allowing the users to quickly formulate the problem at hand and solve it, resulting in an intuitive frameworks designed by researchers for researchers."
The other thing that is missing is a description of the inference workflow for how the final trained model is expected to be used.

**Additional context**
For JOSS review: https://github.com/openjournals/joss-reviews/issues/5352
