+++ pina/problem/abstract_problem.py
@@ -116,9 +116,29 @@ class AbstractProblem(metaclass=ABCMeta):
         Generate a set of points to span the `Location` of all the conditions of
         the problem.
 
-        >>> pinn.span_pts(n=10, mode='grid')
-        >>> pinn.span_pts(n=10, mode='grid', location=['bound1'])
-        >>> pinn.span_pts(n=10, mode='grid', variables=['x'])
+        :param n: Number of points to sample, see Note below
+            for reference.
+        :type n: int
+        :param mode: Mode for sampling, defaults to ``random``.
+            Available modes include: random sampling, ``random``;
+            latin hypercube sampling, ``latin`` or ``lh``;
+            chebyshev sampling, ``chebyshev``; grid sampling ``grid``.
+        :param variables: problem's variables to be sampled, defaults to 'all'.
+        :type variables: str or list[str], optional
+        :param locations: problem's locations from where to sample, defaults to 'all'.
+        :type locations: str, optional
+
+        :Example:
+            >>> pinn.span_pts(n=10, mode='grid')
+            >>> pinn.span_pts(n=10, mode='grid', location=['bound1'])
+            >>> pinn.span_pts(n=10, mode='grid', variables=['x'])
+
+        .. warning::
+            ``random`` is currently the only implemented ``mode`` for all geometries, i.e.
+            ``EllipsoidDomain``, ``CartesianDomain``, ``SimplexDomain`` and the geometries
+            compositions ``Union``, ``Difference``, ``Exclusion``, ``Intersection``. The
+            modes ``latin`` or ``lh``,  ``chebyshev``, ``grid`` are only implemented for
+            ``CartesianDomain``.
         """
 
         # check consistecy n
@@ -152,11 +172,22 @@ class AbstractProblem(metaclass=ABCMeta):
         # sampling
         for location in locations:
             condition = self.conditions[location]
+
             # we try to check if we have already sampled
             try:
                 already_sampled = [self.input_pts[location]]
+            # if we have not sampled, a key error is thrown 
             except KeyError:
                 already_sampled = []
+
+            # if we have already sampled fully the condition
+            # but we want to sample again we set already_sampled
+            # to an empty list since we need to sample again, and
+            # self._have_sampled_points to False.
+            if self._have_sampled_points[location]:
+                already_sampled = []
+                self._have_sampled_points[location] = False
+
             # build samples
             samples = [condition.location.sample(
                             n=n,
@@ -168,7 +199,7 @@ class AbstractProblem(metaclass=ABCMeta):
             # setting the grad
             self.input_pts[location].requires_grad_(True)
             self.input_pts[location].retain_grad()
-            # the condition is sampled
+            # the condition is sampled if input_pts contains all labels
             if sorted(self.input_pts[location].labels) ==  sorted(self.input_variables): 
                 self._have_sampled_points[location] = True
 
