### 2023-06-27 mathLab/PINA#140 @dario-coscia
👋🏻 Thanks for the report. Indeed we missed [Modulus](https://github.com/NVIDIA/modulus-sym) from the references. I saw from the GitHub page that the first release was in May, while we sent the paper on April, so maybe that's why we missed. Nevertheless, thank you for the reference, we will add it.

We are currently working i #134 to update the paper explaining better PINA capabilities, and remaining in the total word-count.
