+++ pina/problem/abstract_problem.py
@@ -1,6 +1,6 @@
 """ Module for AbstractProblem class """
 from abc import ABCMeta, abstractmethod
-from ..utils import merge_tensors
+from ..utils import merge_tensors, check_consistency
 
 
 class AbstractProblem(metaclass=ABCMeta):
@@ -24,7 +24,7 @@ class AbstractProblem(metaclass=ABCMeta):
 
         # put in self.input_pts all the points that we don't need to sample
         self._span_condition_points()
-      
+
     @property
     def input_variables(self):
         """
@@ -111,53 +111,97 @@ class AbstractProblem(metaclass=ABCMeta):
                 continue
             self.input_pts[condition_name] = samples
 
-    def discretise_domain(self, *args, **kwargs):
+    def discretise_domain(self, n, mode = 'random', variables = 'all', locations = 'all'):
         """
         Generate a set of points to span the `Location` of all the conditions of
         the problem.
 
-        >>> pinn.span_pts(n=10, mode='grid')
-        >>> pinn.span_pts(n=10, mode='grid', location=['bound1'])
-        >>> pinn.span_pts(n=10, mode='grid', variables=['x'])
+        :param n: Number of points to sample, see Note below
+            for reference.
+        :type n: int
+        :param mode: Mode for sampling, defaults to ``random``.
+            Available modes include: random sampling, ``random``;
+            latin hypercube sampling, ``latin`` or ``lh``;
+            chebyshev sampling, ``chebyshev``; grid sampling ``grid``.
+        :param variables: problem's variables to be sampled, defaults to 'all'.
+        :type variables: str or list[str], optional
+        :param locations: problem's locations from where to sample, defaults to 'all'.
+        :type locations: str, optional
+
+        :Example:
+            >>> pinn.span_pts(n=10, mode='grid')
+            >>> pinn.span_pts(n=10, mode='grid', location=['bound1'])
+            >>> pinn.span_pts(n=10, mode='grid', variables=['x'])
+
+        .. warning::
+            ``random`` is currently the only implemented ``mode`` for all geometries, i.e.
+            ``EllipsoidDomain``, ``CartesianDomain``, ``SimplexDomain`` and the geometries
+            compositions ``Union``, ``Difference``, ``Exclusion``, ``Intersection``. The
+            modes ``latin`` or ``lh``,  ``chebyshev``, ``grid`` are only implemented for
+            ``CartesianDomain``.
         """
-        if all(key in kwargs for key in ['n', 'mode']):
-            argument = {}
-            argument['n'] = kwargs['n']
-            argument['mode'] = kwargs['mode']
-            argument['variables'] = self.input_variables
-            arguments = [argument]
-        elif any(key in kwargs for key in ['n', 'mode']) and args:
-            raise ValueError("Don't mix args and kwargs")
-        elif isinstance(args[0], int) and isinstance(args[1], str):
-            argument = {}
-            argument['n'] = int(args[0])
-            argument['mode'] = args[1]
-            argument['variables'] = self.input_variables
-            arguments = [argument]
-        elif all(isinstance(arg, dict) for arg in args):
-            arguments = args
+
+        # check consistecy n
+        check_consistency(n, int)
+
+        # check consistency mode
+        check_consistency(mode, str)
+        if mode not in ['random', 'grid', 'lh', 'chebyshev', 'latin']:
+            raise TypeError(f'mode {mode} not valid.')
+
+        # check consistency variables
+        if variables == 'all':
+            variables = self.input_variables
         else:
-            raise RuntimeError
+            check_consistency(variables, str)
 
-        locations = kwargs.get('locations', 'all')
+        if sorted(variables) !=  sorted(self.input_variables):
+            TypeError(f'Wrong variables for sampling. Variables ',
[…]
