### 2023-06-28 5998ac7
fix loss init
