### 2023-06-28 5676e5a
Refacotring Geometries - added OperationInterface - redid test cases - editted Union, Intersect, Exclusion, and Difference to inherit from OperationInterface - simplified Union, Intersect, Exclusion, and Difference
