+++ pina/geometry/union_domain.py
@@ -1,11 +1,12 @@
 import torch
 from .location import Location
+from .operation_interface import OperationInterface
 from ..utils import check_consistency
 from ..label_tensor import LabelTensor
 import random
 
 
-class Union(Location):
+class Union(OperationInterface):
     """ PINA implementation of Unions of Domains."""
 
     def __init__(self, geometries):
@@ -23,37 +24,38 @@ class Union(Location):
             >>> union = GeometryUnion([ellipsoid1, ellipsoid2])
 
         """
-        super().__init__()
+        super().__init__(geometries)
+    #     super().__init__()
 
-        # union checks
-        check_consistency(geometries, Location)
-        self._check_union_dimensions(geometries)
+    #     # union checks
+    #     check_consistency(geometries, Location)
+    #     self._check_union_dimensions(geometries)
 
-        # assign geometries
-        self._geometries = geometries
+    #     # assign geometries
+    #     self._geometries = geometries
 
-    @property
-    def geometries(self):
-        """ 
-        The geometries."""
-        return self._geometries
+    # @property
+    # def geometries(self):
+    #     """
+    #     The geometries."""
+    #     return self._geometries
 
-    @property
-    def variables(self):
-        """
-        Spatial variables.
+    # @property
+    # def variables(self):
+    #     """
+    #     Spatial variables.
 
-        :return: All the spatial variables defined in '__init__()' in order.
-        :rtype: list[str]
-        """
-        all_variables = []
-        seen_variables = set()
-        for geometry in self.geometries:
-            for variable in geometry.variables:
-                if variable not in seen_variables:
-                    all_variables.append(variable)
-                    seen_variables.add(variable)
-        return all_variables
+    #     :return: All the spatial variables defined in '__init__()' in order.
+    #     :rtype: list[str]
+    #     """
+    #     all_variables = []
+    #     seen_variables = set()
+    #     for geometry in self.geometries:
+    #         for variable in geometry.variables:
+    #             if variable not in seen_variables:
+    #                 all_variables.append(variable)
+    #                 seen_variables.add(variable)
+    #     return all_variables
 
     def is_inside(self, point, check_border=False):
         """Check if a point is inside the union domain.
@@ -71,66 +73,66 @@ class Union(Location):
                 return True
         return False
 
-    def sample(self, n, mode='random', variables='all'):
-        """Sample routine.
+    # def sample(self, n, mode='random', variables='all'):
+    #     """Sample routine.
 
-        :param n: Number of points to sample in the shape.
-        :type n: int
-        :param mode: Mode for sampling, defaults to 'random'.
-            Available modes include: random sampling, 'random'.
-        :type mode: str, optional
-        :param variables: pinn variable to be sampled, defaults to 'all'.
-        :type variables: str or list[str], optional
+    #     :param n: Number of points to sample in the shape.
+    #     :type n: int
+    #     :param mode: Mode for sampling, defaults to 'random'.
+    #         Available modes include: random sampling, 'random'.
+    #     :type mode: str, optional
+    #     :param variables: pinn variable to be sampled, defaults to 'all'.
+    #     :type variables: str or list[str], optional
 
-        :Example:
-            # Create two ellipsoid domains
-            >>> ellipsoid1 = EllipsoidDomain({'x': [-1, 1], 'y': [-1, 1]})
-            >>> ellipsoid2 = EllipsoidDomain({'x': [0, 2], 'y': [0, 2]})
+    #     :Example:
+    #         # Create two ellipsoid domains
+    #         >>> ellipsoid1 = EllipsoidDomain({'x': [-1, 1], 'y': [-1, 1]})
+    #         >>> ellipsoid2 = EllipsoidDomain({'x': [0, 2], 'y': [0, 2]})
 
-            # Create a union of the ellipsoid domains
[…]
