+++ pina/geometry/exclusion_domain.py
@@ -4,9 +4,10 @@ from .location import Location
 from ..utils import check_consistency
 from ..label_tensor import LabelTensor
 import random
+from .operation_interface import OperationInterface
 
 
-class Exclusion(Location):
+class Exclusion(OperationInterface):
     """ PINA implementation of Exclusion of Domains."""
 
     def __init__(self, geometries):
@@ -23,37 +24,37 @@ class Exclusion(Location):
             # Create a Exclusion of the ellipsoid domains
             >>> exclusion = Exclusion([ellipsoid1, ellipsoid2])
         """
-        super().__init__()
+        super().__init__(geometries)
 
-        # Exclusion checks
-        check_consistency(geometries, Location)
-        self._check_exclusion_dimesions(geometries)
+        # # Exclusion checks
+        # check_consistency(geometries, Location)
+        # self._check_exclusion_dimesions(geometries)
 
-        # assign geometries
-        self._geometries = geometries
+        # # assign geometries
+        # self._geometries = geometries
 
-    @property
-    def geometries(self):
-        """ 
-        The geometries."""
-        return self._geometries
+    # @property
+    # def geometries(self):
+    #     """
+    #     The geometries."""
+    #     return self._geometries
 
-    @property
-    def variables(self):
-        """
-        Spatial variables.
+    # @property
+    # def variables(self):
+    #     """
+    #     Spatial variables.
 
-        :return: All the spatial variables defined in '__init__()' in order.
-        :rtype: list[str]
-        """
-        all_variables = []
-        seen_variables = set()
-        for geometry in self.geometries:
-            for variable in geometry.variables:
-                if variable not in seen_variables:
-                    all_variables.append(variable)
-                    seen_variables.add(variable)
-        return all_variables
+    #     :return: All the spatial variables defined in '__init__()' in order.
+    #     :rtype: list[str]
+    #     """
+    #     all_variables = []
+    #     seen_variables = set()
+    #     for geometry in self.geometries:
+    #         for variable in geometry.variables:
+    #             if variable not in seen_variables:
+    #                 all_variables.append(variable)
+    #                 seen_variables.add(variable)
+    #     return all_variables
 
     def is_inside(self, point, check_border=False):
         """Check if a point is inside the Exclusion domain.
@@ -70,76 +71,76 @@ class Exclusion(Location):
                 flag += 1
         return flag == 1
 
-    def sample(self, n, mode='random', variables='all'):
-        """Sample routine.
+    # def sample(self, n, mode='random', variables='all'):
+    #     """Sample routine.
 
-        :param n: Number of points to sample in the shape.
-        :type n: int
-        :param mode: Mode for sampling, defaults to 'random'.
-            Available modes include: random sampling, 'random'.
-        :type mode: str, optional
-        :param variables: pinn variable to be sampled, defaults to 'all'.
-        :type variables: str or list[str], optional
+    #     :param n: Number of points to sample in the shape.
+    #     :type n: int
+    #     :param mode: Mode for sampling, defaults to 'random'.
+    #         Available modes include: random sampling, 'random'.
+    #     :type mode: str, optional
+    #     :param variables: pinn variable to be sampled, defaults to 'all'.
+    #     :type variables: str or list[str], optional
 
-        :Example:
-            # Create two Cartesian domains
-            >>> cartesian1 = CartesianDomain({'x': [0, 2], 'y': [0, 2]})
-            >>> cartesian2 = CartesianDomain({'x': [1, 3], 'y': [1, 3]})
+    #     :Example:
+    #         # Create two Cartesian domains
+    #         >>> cartesian1 = CartesianDomain({'x': [0, 2], 'y': [0, 2]})
[…]
