### 2023-06-28 2a0ad01
Enhancing Equations - add init file - add docs - fixing bug System of equation, replace torch.stack with torch.hstack - add tests
