+++ pina/equation/__init__.py
@@ -1,0 +1,12 @@
+__all__ = [
+    'SystemEquation',
+    'Equation',
+    'FixedValue',
+    'FixedGradient',
+    'FixedFlux',
+    'Laplace',
+]
+
+from .equation import *
+from .equation_factory import *
+from .system_equation import *
