+++ joss/paper.bib
@@ -10,6 +10,13 @@
   publisher={Now Publishers, Inc.}
 }
 
+@misc{modulussym,
+  title = {{NVIDIA Modulus}},
+  howpublished = "\url{https://github.com/NVIDIA/modulus-sym}",
+  year = {2023},
+  note = "[Online; accessed 27-April-2023]"
+}
+
 @article{Wang_2005,
 doi = {10.1088/0964-1726/14/1/011},
 url = {https://dx.doi.org/10.1088/0964-1726/14/1/011},
@@ -40,19 +47,19 @@ abstract = {We introduce physics-informed neural networks – neural networks th
 
 @misc{pinns,
   doi = {10.48550/ARXIV.2201.05624},
-  
+
   url = {https://arxiv.org/abs/2201.05624},
-  
+
   author = {Cuomo, Salvatore and di Cola, Vincenzo Schiano and Giampaolo, Fabio and Rozza, Gianluigi and Raissi, Maziar and Piccialli, Francesco},
-  
+
   keywords = {Machine Learning (cs.LG), Artificial Intelligence (cs.AI), Numerical Analysis (math.NA), Data Analysis, Statistics and Probability (physics.data-an), FOS: Computer and information sciences, FOS: Computer and information sciences, FOS: Mathematics, FOS: Mathematics, FOS: Physical sciences, FOS: Physical sciences},
-  
+
   title = {Scientific Machine Learning through Physics-Informed Neural Networks: Where we are and What's next},
-  
+
   publisher = {arXiv},
-  
+
   year = {2022},
-  
+
   copyright = {arXiv.org perpetual, non-exclusive license}
 }
 
@@ -138,41 +145,41 @@ abstract = {In this work we investigate the possibility of using physics-informe
 
 @misc{Markidis,
   doi = {10.48550/ARXIV.2103.09655},
-  
+
   url = {https://arxiv.org/abs/2103.09655},
-  
+
   author = {Markidis, Stefano},
-  
+
   keywords = {Numerical Analysis (math.NA), Distributed, Parallel, and Cluster Computing (cs.DC), Computational Physics (physics.comp-ph), FOS: Mathematics, FOS: Mathematics, FOS: Computer and information sciences, FOS: Computer and information sciences, FOS: Physical sciences, FOS: Physical sciences},
-  
+
   title = {The Old and the New: Can Physics-Informed Deep-Learning Replace Traditional Linear Solvers?},
-  
+
   publisher = {arXiv},
-  
+
   year = {2021},
-  
+
   copyright = {arXiv.org perpetual, non-exclusive license}
 }
 
 @article{Kharazmi_2021,
 	doi = {10.1016/j.cma.2020.113547},
-  
+
 	url = {https://doi.org/10.1016%2Fj.cma.2020.113547},
-  
+
 	year = 2021,
 	month = {feb},
-  
+
 	publisher = {Elsevier {BV}
 },
-  
+
 	volume = {374},
-  
+
 	pages = {113547},
-  
+
 	author = {Ehsan Kharazmi and Zhongqiang Zhang and George E.M. Karniadakis},
-  
+
 	title = {hp-{VPINNs}: Variational physics-informed neural networks with domain decomposition},
-  
+
 	journal = {Computer Methods in Applied Mechanics and Engineering}
 }
 
@@ -192,54 +199,54 @@ abstract = {Fatigue life of a wind turbine main bearing is drastically affected 
 
 @misc{strazdemo,
   doi = {10.48550/ARXIV.2110.13530},
-  
+
   url = {https://arxiv.org/abs/2110.13530},
-  
+
   author = {Demo, Nicola and Strazzullo, Maria and Rozza, Gianluigi},
-  
+
[…]
