### 2023-06-28 1383ad3
Variables in Discretise Domain (#139)
* fix problems discretise_domain
* adding docs, fixing tests
