+++ pina/__init__.py
@@ -10,7 +10,7 @@ __all__ = [
 
 from .meta import *
 from .label_tensor import LabelTensor
-from .pinn import PINN
+from .solvers.pinn import PINN
 from .trainer import Trainer
 from .plotter import Plotter
 from .condition import Condition
