+++ pina/solvers/pinn.py
@@ -8,10 +8,9 @@ except ImportError:
 from torch.optim.lr_scheduler import ConstantLR
 
 from .solver import SolverInterface
-from .label_tensor import LabelTensor
-from .utils import check_consistency
-from .writer import Writer
-from .loss import LossInterface
+from ..label_tensor import LabelTensor
+from ..utils import check_consistency
+from ..loss import LossInterface
 from torch.nn.modules.loss import _Loss
 
 
@@ -19,7 +18,18 @@ torch.pi = torch.acos(torch.zeros(1)).item() * 2  # which is 3.1415927410125732
 
 
 class PINN(SolverInterface):
+    """
+    PINN solver class. This class implements Physics Informed Neural 
+    Network solvers, using a user specified ``model`` to solve a specific
+    ``problem``. 
 
+    .. seealso::
+
+        **Original reference**: Karniadakis, G. E., Kevrekidis, I. G., Lu, L., 
+        Perdikaris, P., Wang, S., & Yang, L. (2021). 
+        Physics-informed machine learning. Nature Reviews Physics, 3(6), 422-440.
+        <https://doi.org/10.1038/s42254-021-00314-5>`_.
+    """
     def __init__(self,
                  problem,
                  model,
@@ -45,20 +55,21 @@ class PINN(SolverInterface):
             rate scheduler.
         :param dict scheduler_kwargs: LR scheduler constructor keyword args.
         '''
-        super().__init__(model=model, problem=problem, extra_features=extra_features)
+        super().__init__(models=[model],
+                         problem=problem,
+                         optimizers=[optimizer],
+                         optimizers_kwargs=[optimizer_kwargs],
+                         extra_features=extra_features)
         
         # check consistency 
-        check_consistency(optimizer, torch.optim.Optimizer, subclass=True)
-        check_consistency(optimizer_kwargs, dict)
         check_consistency(scheduler, LRScheduler, subclass=True)
         check_consistency(scheduler_kwargs, dict)
         check_consistency(loss, (LossInterface, _Loss), subclass=False)
 
         # assign variables
-        self._optimizer = optimizer(self.model.parameters(), **optimizer_kwargs)
-        self._scheduler = scheduler(self._optimizer, **scheduler_kwargs)
+        self._scheduler = scheduler(self.optimizers[0], **scheduler_kwargs)
         self._loss = loss
-        self._writer = Writer()
+        self._neural_net = self.models[0]
 
 
     def forward(self, x):
@@ -72,7 +83,7 @@ class PINN(SolverInterface):
         # extract labels
         x = x.extract(self.problem.input_variables)
         # perform forward pass
-        output = self.model(x).as_subclass(LabelTensor)
+        output = self.neural_net(x).as_subclass(LabelTensor)
         # set the labels
         output.labels = self.problem.output_variables
         return output
@@ -84,7 +95,7 @@ class PINN(SolverInterface):
         :return: The optimizers and the schedulers
         :rtype: tuple(list, list)
         """
-        return [self._optimizer], [self._scheduler]
+        return self.optimizers, [self.scheduler]
     
     def training_step(self, batch, batch_idx):
         """PINN solver training step.
@@ -109,11 +120,11 @@ class PINN(SolverInterface):
             # PINN loss: equation evaluated on location or input_points
             if hasattr(condition, 'equation'):
                 target = condition.equation.residual(samples, self.forward(samples))
-                loss = self._loss(torch.zeros_like(target), target)
+                loss = self.loss(torch.zeros_like(target), target)
             # PINN loss: evaluate model(input_points) vs output_points
             elif hasattr(condition, 'output_points'):
                 input_pts, output_pts = samples
-                loss = self._loss(self.forward(input_pts), output_pts)
[…]
