+++ pina/solvers/__init__.py
@@ -1,0 +1,7 @@
+__all__ = [
+    'PINN',
+    'GAROM',
+]
+
+from .garom import GAROM
+from .pinn import PINN
