### 2023-06-28 be6a7a0
Update test_pinn.py
