### 2023-06-28 a42aea5
device for sample points in absProblem (#132)
* device for sample points in absProblem
