### 2023-06-28 77e8172
device for sample points in absProblem (#132)
* device for sample points in absProblem
