### 2023-06-28 b6adcff
Issue #95 fixed and improved plot_loss (#106)
* plotter.py plot_loss fixed Issue #95
* Improved the plot_loss to show title and labels.
