+++ joss/paper.bib
@@ -12,7 +12,7 @@
 
 @misc{modulussym,
   title = {{NVIDIA Modulus}},
-  howpublished = "\url{https://github.com/NVIDIA/modulus-sym}",
+  howpublished = "\url{https://github.com/NVIDIA/modulus}",
   year = {2023},
   note = "[Online; accessed 27-April-2023]"
 }
