### 2023-06-28 mathLab/PINA#140 @ndem0
Dear @akshaysubr, we extended the paper with more information about the pros of PINA, highlighting why the statement "easiness of usage".
We did not put any lines of code or example directly in the final paper, limiting to insert the links to such examples and the documentation, not only for space reasons but also for possible future changes in the source code (which would make the PDF version outdated).
Please let us know if we forgot some corrections!

PS. We add reference to the entire Modulus package, and not only to the modulus-syn
