+++ pina/geometry/intersection_domain.py
@@ -42,63 +42,61 @@ class Intersection(OperationInterface):
                 flag += 1
         return flag == len(self.geometries)
 
-    # def sample(self, n, mode='random', variables='all'):
-    #     """Sample routine.
+    def sample(self, n, mode='random', variables='all'):
+        """Sample routine for intersection domain.
 
-    #     :param n: Number of points to sample in the shape.
-    #     :type n: int
-    #     :param mode: Mode for sampling, defaults to 'random'.
-    #         Available modes include: random sampling, 'random'.
-    #     :type mode: str, optional
-    #     :param variables: pinn variable to be sampled, defaults to 'all'.
-    #     :type variables: str or list[str], optional
+        :param n: Number of points to sample in the shape.
+        :type n: int
+        :param mode: Mode for sampling, defaults to 'random'.
+            Available modes include: random sampling, 'random'.
+        :type mode: str, optional
+        :param variables: pinn variable to be sampled, defaults to 'all'.
+        :type variables: str or list[str], optional
 
-    #     :Example:
-    #         # Create two Cartesian domains
-    #         >>> cartesian1 = CartesianDomain({'x': [0, 2], 'y': [0, 2]})
-    #         >>> cartesian2 = CartesianDomain({'x': [1, 3], 'y': [1, 3]})
+        :Example:
+            # Create two Cartesian domains
+            >>> cartesian1 = CartesianDomain({'x': [0, 2], 'y': [0, 2]})
+            >>> cartesian2 = CartesianDomain({'x': [1, 3], 'y': [1, 3]})
 
-    #         # Create a Intersection of the ellipsoid domains
-    #         >>> intersection = Intersection([cartesian1, cartesian2])
+            # Create a Intersection of the ellipsoid domains
+            >>> intersection = Intersection([cartesian1, cartesian2])
 
-    #         >>> intersection.sample(n=1000)
-    #             LabelTensor([[1.2389, 1.4923],
-    #                         [1.7237, 1.8469],
-    #                         [1.5058, 1.4860],
-    #                         ...,
-    #                         [1.4470, 1.4379],
-    #                         [1.0176, 1.9371],
-    #                         [1.4728, 1.8302]])
+            >>> intersection.sample(n=5)
+                LabelTensor([[1.7697, 1.8654],
+                            [1.2841, 1.1208],
+                            [1.7289, 1.9843],
+                            [1.3332, 1.2448],
+                            [1.9902, 1.4458]])
 
-    #         >>> len(intersection.sample(n=1000)
-    #             1000
+            >>> len(intersection.sample(n=5)
+                5
 
-    #     """
-    #     if mode != 'random':
-    #         raise NotImplementedError(
-    #             f'{mode} is not a valid mode for sampling.')
+        """
+        if mode != 'random':
+            raise NotImplementedError(
+                f'{mode} is not a valid mode for sampling.')
 
-    #     sampled = []
+        sampled = []
 
-    #     # calculate the number of points to sample for each geometry and the remainder.
-    #     remainder = n % len(self.geometries)
-    #     num_points = n // len(self.geometries)
+        # calculate the number of points to sample for each geometry and the remainder.
+        remainder = n % len(self.geometries)
+        num_points = n // len(self.geometries)
 
-    #     # sample the points
-    #     # NB. geometries as shuffled since if we sample
-    #     # multiple times just one point, we would end
-    #     # up sampling only from the first geometry.
-    #     iter_ = random.sample(self.geometries, len(self.geometries))
-    #     for i, geometry in enumerate(iter_):
-    #         sampled_points = []
-    #         # int(i < remainder) is one only if we have a remainder
-    #         # different than zero. Notice that len(geometries) is
-    #         # always smaller than remaider.
[…]
