+++ tests/test_geometry/test_union.py
@@ -50,11 +50,11 @@ def test_sample():
     n = 100
     domain = Union([EllipsoidDomain({'x': [-1, 1], 'y': [-1, 1]}),
                     CartesianDomain({'x': [-0.5, 0.5], 'y': [-0.5, 0.5]})])
-    pts = domain.sample(n, type="union")
+    pts = domain.sample(n)
     assert isinstance(pts, LabelTensor)
     assert pts.shape[0] == n
 
     n = 105
-    pts = domain.sample(n, type="union")
+    pts = domain.sample(n)
     assert isinstance(pts, LabelTensor)
     assert pts.shape[0] == n
