+++ tests/test_solvers/test_pinn.py
@@ -96,6 +96,18 @@ def test_train_cpu():
     trainer = Trainer(solver=pinn, kwargs={'max_epochs' : 5, 'accelerator':'cpu'})
     trainer.train()
 
+def test_train_cpu_sampling_few_vars():
+    poisson_problem = Poisson()
+    boundaries = ['gamma1', 'gamma2', 'gamma3', 'gamma4']
+    n = 10
+    poisson_problem.discretise_domain(n, 'grid', locations=boundaries)
+    poisson_problem.discretise_domain(n, 'random', locations=['D'], variables=['x'])
+    poisson_problem.discretise_domain(n, 'random', locations=['D'], variables=['y'])
+    pinn = PINN(problem = poisson_problem, model=model, extra_features=None, loss=LpLoss())
+    trainer = Trainer(solver=pinn, kwargs={'max_epochs' : 5, 'accelerator':'cpu'})
+    trainer.train()
+
+
 def test_train_extra_feats_cpu():
     poisson_problem = Poisson()
     boundaries = ['gamma1', 'gamma2', 'gamma3', 'gamma4']
