### 2023-06-29 29b586a
fix paper
changing some lines to make it more general for future version of the code.
