### 2023-06-29 3100e06
Simplex Domain * much cleaner code * is_inside much less complicated * tests pass
