### 2023-06-29 3e1902a
Simplex Domain * much cleaner code * is_inside much less complicated * tests pass
