+++ pina/geometry/simplex.py
@@ -5,6 +5,7 @@ from .location import Location
 from pina import LabelTensor
 from ..utils import check_consistency
 
+
 class SimplexDomain(Location):
     """PINA implementation of a Simplex."""
 
@@ -21,7 +22,13 @@ class SimplexDomain(Location):
             frontier are taken. If `sample_surface=False`, no such criteria
             is followed.
         :type sample_surface: bool
-        
+
+        .. warning::
+            Sampling for dimensions greater or equal to 10 could result
+            in a shrinking of the simplex, which degrades the quality
+            of the samples. For dimensions higher than 10, other algorithms
+            for sampling should be used.
+
         :Example:
 
             >>> spatial_domain = SimplexDomain({'vertex1': [0, 0], 'vertex2': [1, 1], 'vertex3': [0, 2]}, ['x', 'y'])
@@ -30,10 +37,9 @@ class SimplexDomain(Location):
         # check consistency of labels
         if not isinstance(labels, list):
             raise ValueError(f"{type(labels).__name__} must be {list}.")
-        for lab in labels:
-            check_consistency(lab, str)
+        check_consistency(labels, str)
         self._coordinate_labels = labels
-        
+
         # check consistency of sample_surface
         check_consistency(sample_surface, bool)
         self._sample_surface = sample_surface
@@ -43,17 +49,18 @@ class SimplexDomain(Location):
         for vertex in simplex_dict.values():
             if not isinstance(vertex, list):
                 raise ValueError(f"{type(vertex).__name__} must be {list}.")
-        
 
-        # vertices, vectors, dimension 
-        self._vertices = {label: LabelTensor(torch.tensor([vertex]), self.variables) for label, vertex in simplex_dict.items()}
-        self._vectors = self._basis_vectors(self._vertices_list)
-        self._dimension = len(simplex_dict)-1
-
+        # vertices, vectors, dimension
+        self._vertices = simplex_dict
+        self._vert_list = torch.tensor(list(simplex_dict.values())).T
+        self._vectors = (
+            self._vertices_list[:, :-1] - self._vertices_list[:, None, -1]
+        ).type(torch.FloatTensor)
 
         # build cartesian_bound
-        self._cartesian_bound = self._build_cartesian(list(simplex_dict.values()), labels)
-
+        self._cartesian_bound = self._build_cartesian(
+            list(simplex_dict.values()), labels
+        )
 
     @property
     def variables(self):
@@ -66,7 +73,6 @@ class SimplexDomain(Location):
 
         return self._coordinate_labels
 
-
     @property
     def vertices(self):
         """
@@ -77,7 +83,6 @@ class SimplexDomain(Location):
         """
 
         return self._vertices
-    
 
     @property
     def vectors(self):
@@ -88,7 +93,6 @@ class SimplexDomain(Location):
         :rtype: dict(LabelTensor)
         """
         return self._vectors
-    
 
     @property
     def _vertices_list(self):
@@ -98,8 +102,7 @@ class SimplexDomain(Location):
         :return: List of vectors
         :rtype: list[LabelTensor]
         """
-        return list(self.vertices.values())
-    
+        return self._vert_list
 
     @property
     def cartesian_bound(self):
@@ -111,19 +114,6 @@ class SimplexDomain(Location):
         """
 
         return self._cartesian_bound
-    
-
-    @property
-    def dimension(self):
-        """
-        Dimension of Simplex domain.
-        
-        :return: dimension
-        :rtype: int
-        """
-
-        return self._dimension
-    
 
     @property
     def sample_surface(self):
@@ -135,25 +125,6 @@ class SimplexDomain(Location):
         """
 
         return self._sample_surface
-    
-
-    def _basis_vectors(self, vertices):
-        """
-        Basis vectors for simplex.
-
[…]
