+++ pina/__init__.py
@@ -5,8 +5,7 @@ __all__ = [
     'Plotter',
     'Condition',
     'Location',
-    'CartesianDomain',
-    'SimplexDomain'
+    'CartesianDomain'
 ]
 
 from .meta import *
@@ -17,4 +16,3 @@ from .plotter import Plotter
 from .condition import Condition
 from .geometry import Location
 from .geometry import CartesianDomain
-from .geometry import SimplexDomain
