+++ pina/geometry/simplex.py
@@ -1,7 +1,6 @@
 import torch
-
-from pina.geometry import CartesianDomain
 from .location import Location
+from pina.geometry import CartesianDomain
 from pina import LabelTensor
 from ..utils import check_consistency
 
@@ -12,10 +11,11 @@ class SimplexDomain(Location):
     def __init__(self, simplex_dict, labels, sample_surface=False):
         """
         :param simplex_dict: A dictionary with dict-key a string representing
-            the input variables for the pinn, and dict-value a list
+            the input variables for the problem, and dict-value a list
             representing vertices of the simplex.
         :type simplex_dict: dict
-        :param labels: A list of labels for vertex components
+        :param labels: A list of labels for vertex components. Represents the
+            order in which points should list coordinates.
         :type labels: list[str]
         :param sample_surface: A variable for choosing sample strategies. If
             `sample_surface=True` only samples on the Simplex surface
@@ -38,7 +38,7 @@ class SimplexDomain(Location):
         if not isinstance(labels, list):
             raise ValueError(f"{type(labels).__name__} must be {list}.")
         check_consistency(labels, str)
-        self._coordinate_labels = labels
+        self._labels = labels
 
         # check consistency of sample_surface
         check_consistency(sample_surface, bool)
@@ -52,9 +52,9 @@ class SimplexDomain(Location):
 
         # vertices, vectors, dimension
         self._vertices = simplex_dict
-        self._vert_list = torch.tensor(list(simplex_dict.values())).T
+        self._vertices_matrix = torch.tensor(list(simplex_dict.values())).T
         self._vectors = (
-            self._vertices_list[:, :-1] - self._vertices_list[:, None, -1]
+            self._vertices_matrix[:, :-1] - self._vertices_matrix[:, None, -1]
         ).type(torch.FloatTensor)
 
         # build cartesian_bound
@@ -64,67 +64,7 @@ class SimplexDomain(Location):
 
     @property
     def variables(self):
-        """
-        Coordinate labels of simplex.
-
-        :return: Coordinate labels
-        :rtype: list[str]
-        """
-
-        return self._coordinate_labels
-
-    @property
-    def vertices(self):
-        """
-        Vertices of simplex.
-
-        :return: Vectors
-        :rtype: list[list]
-        """
-
         return self._vertices
-
-    @property
-    def vectors(self):
-        """
-        Vectors.
-
-        :return: Vectors
-        :rtype: dict(LabelTensor)
-        """
-        return self._vectors
-
-    @property
-    def _vertices_list(self):
-        """
-        List of vectors.
-
-        :return: List of vectors
-        :rtype: list[LabelTensor]
-        """
-        return self._vert_list
-
-    @property
-    def cartesian_bound(self):
-        """
-        Cartesian border for Simplex domain.
-
-        :return: Cartesian border for Simplex domain
-        :rtype: CartesianDomain
-        """
-
-        return self._cartesian_bound
-
-    @property
-    def sample_surface(self):
-        """
-        Whether the surface should be sampled or not.
-
-        :return: Whether the surface should be sampled or not
-        :rtype: bool
-        """
-
-        return self._sample_surface
 
     def _build_cartesian(self, vertices, labels):
         """
@@ -149,8 +89,11 @@ class SimplexDomain(Location):
     def is_inside(self, point, check_border=False):
         """
         Check if a point is inside the simplex.
-        Uses the algorithm described here: 
+        Uses the algorithm described here:
         https://math.stackexchange.com/questions/1226707/how-to-check-if-point-x-in-mathbbrn-is-in-a-n-simplex
+
+        Barycentric coordinates are also useful:
+        https://en.wikipedia.org/wiki/Barycentric_coordinate_system
 
[…]
