### 2023-06-29 44cf800
add callable function as reduction
