### 2023-06-29 1050b01
minor fixes
