### 2023-06-29 dbd8424
minor fixes
