+++ tests/test_simplex.py
@@ -5,12 +5,34 @@ from pina import LabelTensor
 from pina.geometry import SimplexDomain
 
 def test_constructor():
-    SimplexDomain({'vertex1': [0, 0], 'vertex2': [1, 1], 'vertex3': [2, 0]}, labels=['x', 'y'])
-    SimplexDomain({'vertex1': [1, 2, 3], 'vertex2': [2, 2, 3], 'vertex3': [1, 3, 3], 'vertex4': [1, 2, 9]}, labels=['x', 'y', 'z'])
+    SimplexDomain({'vertex1': [0, 0],
+                   'vertex2': [1, 1],
+                   'vertex3': [2, 0]},
+                   labels=['x', 'y'])
+    SimplexDomain({'vertex1': [0, 0],
+                   'vertex2': [1, 1],
+                   'vertex3': [2, 0]},
+                   labels=['x', 'y'], sample_surface=True)
+    with pytest.raises(ValueError):
+        SimplexDomain({'vertex1': [0, 0],
+                'vertex2': [1, 1],
+                'vertex3': [2, 0]},
+                labels='x')
+        SimplexDomain({'vertex1': [0, 0],
+                'vertex2': [1, 1],
+                'vertex3': [2, 0]},
+                labels=[1, 2])
+        SimplexDomain([3, 2, 1],
+                labels=['x', 'y'])
+
 
 def test_is_inside_2D_check_border_true():
-    domain = SimplexDomain({'vertex1': [0, 0], 'vertex2': [2, 2], 'vertex3': [4, 0]}, ['x', 'y'], sample_surface=True)
-    pt1 = LabelTensor(torch.tensor([[0, 0]]), ['x', 'y']) # TODO: ZERO BUG
+    domain = SimplexDomain({'vertex1': [0, 0], 
+                            'vertex2': [2, 2],
+                            'vertex3': [4, 0]},
+                            ['x', 'y'],
+                            sample_surface=True)
+    pt1 = LabelTensor(torch.tensor([[0, 0]]), ['x', 'y']) 
     pt2 = LabelTensor(torch.tensor([[2, 2]]), ['x', 'y'])
     pt3 = LabelTensor(torch.tensor([[4, 0]]), ['x', 'y'])
     pt4 = LabelTensor(torch.tensor([[3, 1]]), ['x', 'y'])
@@ -21,8 +43,12 @@ def test_is_inside_2D_check_border_true():
         assert domain.is_inside(point=pt, check_border=True) == exp_result
 
 def test_is_inside_2D_check_border_false():
-    domain = SimplexDomain({'vertex1': [0, 0], 'vertex2': [2, 2], 'vertex3': [4, 0]}, ['x', 'y'], sample_surface=True)
-    pt1 = LabelTensor(torch.tensor([[0, 0]]), ['x', 'y']) # TODO: ZERO BUG
+    domain = SimplexDomain({'vertex1': [0, 0],
+                            'vertex2': [2, 2],
+                            'vertex3': [4, 0]}, 
+                            ['x', 'y'], 
+                            sample_surface=True)
+    pt1 = LabelTensor(torch.tensor([[0, 0]]), ['x', 'y']) 
     pt2 = LabelTensor(torch.tensor([[2, 2]]), ['x', 'y'])
     pt3 = LabelTensor(torch.tensor([[4, 0]]), ['x', 'y'])
     pt4 = LabelTensor(torch.tensor([[3, 1]]), ['x', 'y'])
[…]
