+++ tests/test_simplex.py
@@ -4,10 +4,7 @@ import pytest
 from pina import LabelTensor
 from pina.geometry import SimplexDomain
 
-<<<<<<< HEAD
 
-=======
->>>>>>> 3e1902a (Simplex Domain)
 def test_constructor():
     SimplexDomain(
         [
@@ -42,7 +39,6 @@ def test_constructor():
 
 
 def test_is_inside_2D_check_border_true():
-<<<<<<< HEAD
     domain = SimplexDomain(
         [
             LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
@@ -57,22 +53,12 @@ def test_is_inside_2D_check_border_true():
     pt4 = LabelTensor(torch.tensor([[3, 1]]), ["x", "y"])
     pt5 = LabelTensor(torch.tensor([[2, 1]]), ["x", "y"])
     pt6 = LabelTensor(torch.tensor([[100, 100]]), ["x", "y"])
-=======
-    domain = SimplexDomain({'vertex1': [0, 0], 'vertex2': [2, 2], 'vertex3': [4, 0]}, ['x', 'y'], sample_surface=True)
-    pt1 = LabelTensor(torch.tensor([[0, 0]]), ['x', 'y']) # TODO: ZERO BUG
-    pt2 = LabelTensor(torch.tensor([[2, 2]]), ['x', 'y'])
-    pt3 = LabelTensor(torch.tensor([[4, 0]]), ['x', 'y'])
-    pt4 = LabelTensor(torch.tensor([[3, 1]]), ['x', 'y'])
-    pt5 = LabelTensor(torch.tensor([[2, 1]]), ['x', 'y'])
-    pt6 = LabelTensor(torch.tensor([[100, 100]]), ['x', 'y'])
->>>>>>> 3e1902a (Simplex Domain)
     pts = [pt1, pt2, pt3, pt4, pt5, pt6]
     for pt, exp_result in zip(pts, [True, True, True, True, False, False]):
         assert domain.is_inside(point=pt, check_border=True) == exp_result
 
 
 def test_is_inside_2D_check_border_false():
-<<<<<<< HEAD
     domain = SimplexDomain(
         [
             LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
@@ -88,23 +74,12 @@ def test_is_inside_2D_check_border_false():
     pt5 = LabelTensor(torch.tensor([[2, 1]]), ["x", "y"])
     pt6 = LabelTensor(torch.tensor([[2.5, 1]]), ["x", "y"])
     pt7 = LabelTensor(torch.tensor([[100, 100]]), ["x", "y"])
-=======
-    domain = SimplexDomain({'vertex1': [0, 0], 'vertex2': [2, 2], 'vertex3': [4, 0]}, ['x', 'y'], sample_surface=True)
-    pt1 = LabelTensor(torch.tensor([[0, 0]]), ['x', 'y']) # TODO: ZERO BUG
-    pt2 = LabelTensor(torch.tensor([[2, 2]]), ['x', 'y'])
-    pt3 = LabelTensor(torch.tensor([[4, 0]]), ['x', 'y'])
-    pt4 = LabelTensor(torch.tensor([[3, 1]]), ['x', 'y'])
-    pt5 = LabelTensor(torch.tensor([[2, 1]]), ['x', 'y'])
-    pt6 = LabelTensor(torch.tensor([[2.5, 1]]), ['x', 'y'])
-    pt7 = LabelTensor(torch.tensor([[100, 100]]), ['x', 'y'])
->>>>>>> 3e1902a (Simplex Domain)
     pts = [pt1, pt2, pt3, pt4, pt5, pt6, pt7]
[…]
