### 2023-07-04 2bdd7fc
Update paper.md (#134)
* update paper.md and paper.bib
* correct modulus reference

---------

Co-authored-by: Anna Ivagnes <s274001@studenti.polito.it>
Co-authored-by: Dario Coscia <93731561+dario-coscia@users.noreply.github.com>
