+++ .github/workflows/testing_pr.yml
@@ -4,6 +4,7 @@ on:
   pull_request:
     branches:
       - "master"
+      - "v0.1"
 
 
 jobs:
