### 2023-07-04 3187bce
fix doc/readme/joss (#146)
* fix doc/readme/joss with API scheme

---------

Co-authored-by: Dario Coscia <dariocoscia@cli-10-110-11-236.WIFIeduroamSTUD.units.it>
