+++ pina/geometry/exclusion_domain.py
@@ -1,16 +1,17 @@
+"""Module for Location class."""
 import torch
 from .location import Location
-from .operation_interface import OperationInterface
 from ..utils import check_consistency
 from ..label_tensor import LabelTensor
 import random
+from .operation_interface import OperationInterface
 
 
-class Union(OperationInterface):
-    """ PINA implementation of Unions of Domains."""
+class Exclusion(OperationInterface):
+    """ PINA implementation of Exclusion of Domains."""
 
     def __init__(self, geometries):
-        """ PINA implementation of Unions of Domains.
+        """ PINA implementation of Exclusion of Domains.
 
         :param list geometries: A list of geometries from 'pina.geometry' 
             such as 'EllipsoidDomain' or 'CartesianDomain'.
@@ -20,30 +21,28 @@ class Union(OperationInterface):
             >>> ellipsoid1 = EllipsoidDomain({'x': [-1, 1], 'y': [-1, 1]})
             >>> ellipsoid2 = EllipsoidDomain({'x': [0, 2], 'y': [0, 2]})
 
-            # Create a union of the ellipsoid domains
-            >>> union = GeometryUnion([ellipsoid1, ellipsoid2])
-
+            # Create a Exclusion of the ellipsoid domains
+            >>> exclusion = Exclusion([ellipsoid1, ellipsoid2])
         """
         super().__init__(geometries)
 
     def is_inside(self, point, check_border=False):
-        """Check if a point is inside the union domain.
+        """Check if a point is inside the Exclusion domain.
 
         :param point: Point to be checked.
-        :type point: LabelTensor
-        :param check_border: Check if the point is also on the frontier
-            of the ellipsoid, default False.
-        :type check_border: bool
-        :return: Returning True if the point is inside, False otherwise.
+        :type point: torch.Tensor   
+        :param bool check_border: If True, the border is considered inside.
+        :return: True if the point is inside the Exclusion domain, False otherwise.
         :rtype: bool
         """
+        flag = 0
         for geometry in self.geometries:
             if geometry.is_inside(point, check_border):
-                return True
-        return False
+                flag += 1
+        return flag == 1
 
     def sample(self, n, mode='random', variables='all'):
-        """Sample routine for union domain.
+        """Sample routine for exclusion domain.
 
         :param n: Number of points to sample in the shape.
         :type n: int
@@ -54,26 +53,31 @@ class Union(OperationInterface):
         :type variables: str or list[str], optional
 
         :Example:
-            # Create two ellipsoid domains
+            # Create two Cartesian domains
             >>> cartesian1 = CartesianDomain({'x': [0, 2], 'y': [0, 2]})
             >>> cartesian2 = CartesianDomain({'x': [1, 3], 'y': [1, 3]})
 
-            # Create a union of the ellipsoid domains
-            >>> union = Union([cartesian1, cartesian2])
+            # Create a Exclusion of the ellipsoid domains
+            >>> Exclusion = Exclusion([cartesian1, cartesian2])
 
-            >>> union.sample(n=5)
-                LabelTensor([[1.2128, 2.1991],
-                            [1.3530, 2.4317],
-                            [2.2562, 1.6605],
-                            [0.8451, 1.9878],
-                            [1.8623, 0.7102]])
+            >>> Exclusion.sample(n=5)
+                LabelTensor([[2.4187, 1.5792],
+                            [2.7456, 2.3868],
+                            [2.3830, 1.7037],
+                            [0.8636, 1.8453],
+                            [0.1978, 0.3526]])
 
-            >>> len(union.sample(n=5)
+            >>> len(Exclusion.sample(n=5)
                 5
-        """
-        sampled_points = []
 
-        # calculate the number of points to sample for each geometry and the remainder
+        """
+        if mode != 'random':
+            raise NotImplementedError(
[…]
