+++ pina/geometry/__init__.py
@@ -2,11 +2,18 @@ __all__ = [
     'Location',
     'CartesianDomain',
     'EllipsoidDomain',
-    'Union'
+    'Union',
+    'Intersection',
+    'Exclusion',
+    'Difference',
+    'OperationInterface'
 ]
 
 from .location import Location
 from .cartesian import CartesianDomain
 from .ellipsoid import EllipsoidDomain
-from .difference_domain import Difference
+from .exclusion_domain import Exclusion
+from .intersection_domain import Intersection
 from .union_domain import Union
+from .difference_domain import Difference
+from .operation_interface import OperationInterface
