### 2023-07-05 c0cad8c
SimplexDomain object now takes list of LabelTensor objects as only mandatory parameter, each representing a vertex of the simplex
