+++ tests/test_simplex.py
@@ -4,96 +4,124 @@ import pytest
 from pina import LabelTensor
 from pina.geometry import SimplexDomain
 
+
 def test_constructor():
-    SimplexDomain({'vertex1': [0, 0],
-                   'vertex2': [1, 1],
-                   'vertex3': [2, 0]},
-                   labels=['x', 'y'])
-    SimplexDomain({'vertex1': [0, 0],
-                   'vertex2': [1, 1],
-                   'vertex3': [2, 0]},
-                   labels=['x', 'y'], sample_surface=True)
+    SimplexDomain(
+        [
+            LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[1, 1]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+        ]
+    )
+    SimplexDomain(
+        [
+            LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[1, 1]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+        ],
+        sample_surface=True,
+    )
     with pytest.raises(ValueError):
-        SimplexDomain({'vertex1': [0, 0],
-                'vertex2': [1, 1],
-                'vertex3': [2, 0]},
-                labels='x')
-        SimplexDomain({'vertex1': [0, 0],
-                'vertex2': [1, 1],
-                'vertex3': [2, 0]},
-                labels=[1, 2])
-        SimplexDomain([3, 2, 1],
-                labels=['x', 'y'])
+        SimplexDomain(
+            [
+                LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+                LabelTensor(torch.tensor([[1, 1]]), labels=["x", "z"]),
+                LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+            ]
+        )
+        SimplexDomain(
+            [
+                LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+                [1, 1],
+                LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+            ]
+        )
 
 
 def test_is_inside_2D_check_border_true():
-    domain = SimplexDomain({'vertex1': [0, 0], 
-                            'vertex2': [2, 2],
-                            'vertex3': [4, 0]},
-                            ['x', 'y'],
-                            sample_surface=True)
-    pt1 = LabelTensor(torch.tensor([[0, 0]]), ['x', 'y']) 
-    pt2 = LabelTensor(torch.tensor([[2, 2]]), ['x', 'y'])
-    pt3 = LabelTensor(torch.tensor([[4, 0]]), ['x', 'y'])
-    pt4 = LabelTensor(torch.tensor([[3, 1]]), ['x', 'y'])
-    pt5 = LabelTensor(torch.tensor([[2, 1]]), ['x', 'y'])
-    pt6 = LabelTensor(torch.tensor([[100, 100]]), ['x', 'y'])
+    domain = SimplexDomain(
+        [
+            LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
[…]
