+++ pina/geometry/simplex.py
@@ -8,15 +8,12 @@ from ..utils import check_consistency
 class SimplexDomain(Location):
     """PINA implementation of a Simplex."""
 
-    def __init__(self, simplex_dict, labels, sample_surface=False):
+    def __init__(self, simplex_matrix, sample_surface=False):
         """
-        :param simplex_dict: A dictionary with dict-key a string representing
-            the input variables for the problem, and dict-value a list
-            representing vertices of the simplex.
-        :type simplex_dict: dict
-        :param labels: A list of labels for vertex components. Represents the
-            order in which points should list coordinates.
-        :type labels: list[str]
+        :param simplex_matrix: A matrix of LabelTensor objects representing
+            a vertex of the simplex (a tensor), and the coordinates of the
+            point (a list of labels).
+        :type simplex_matrix: list[LabelTensor]
         :param sample_surface: A variable for choosing sample strategies. If
             `sample_surface=True` only samples on the Simplex surface
             frontier are taken. If `sample_surface=False`, no such criteria
@@ -31,46 +28,63 @@ class SimplexDomain(Location):
 
         :Example:
 
-            >>> spatial_domain = SimplexDomain({'vertex1': [0, 0], 
-                                                'vertex2': [1, 1], 
-                                                'vertex3': [0, 2]}, 
-                                                ['x', 'y'])
+            >>> spatial_domain = SimplexDomain(
+                    [
+                        LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+                        LabelTensor(torch.tensor([[1, 1]]), labels=["x", "y"]),
+                        LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+                    ], sample_surface = True
+                )
         """
-
-        # check consistency of labels
-        if not isinstance(labels, list):
-            raise ValueError(f"{type(labels).__name__} must be {list}.")
-        check_consistency(labels, str)
-        self._variables = labels
 
         # check consistency of sample_surface
         check_consistency(sample_surface, bool)
         self._sample_surface = sample_surface
 
         # check consistency of simplex_dict
-        check_consistency(simplex_dict, dict)
-        for vertex in simplex_dict.values():
-            if not isinstance(vertex, list):
-                raise ValueError(f"{type(vertex).__name__} must be {list}.")
+        if not isinstance(simplex_matrix, list):
+            raise ValueError(f"{type(vertex).__name__} must be {list}.")
+        try:
+            _labels = simplex_matrix[0].labels
+        except:
+            raise ValueError(
+                f"{type(simplex_matrix[0]).__name__} must be {LabelTensor}."
+            )
 
-        # vertices, vectors, dimension
-        self._vertices_matrix = torch.tensor(list(simplex_dict.values()), dtype=torch.float).T
-        self._vectors_shifted = self._vertices_matrix[:, :-1] - self._vertices_matrix[:, None, -1]
+        for vertex in simplex_matrix:
+            if not isinstance(vertex, LabelTensor):
+                raise ValueError(f"{type(vertex).__name__} must be {LabelTensor}.")
+            if vertex.labels != _labels:
+                raise ValueError(f"Labels don't match.")
+
+        # vertices, vectors, coordinates
+        self._vertices_matrix = torch.tensor(
+            [
+                [float(vertex.extract(label)) for label in vertex.labels]
+                for vertex in simplex_matrix
+            ]
+        ).T
+        self._vectors_shifted = (
+            self._vertices_matrix[:, :-1] - self._vertices_matrix[:, None, -1]
+        )
+        self._coordinates = simplex_matrix[0].labels
 
         # build cartesian_bound
-        self._cartesian_bound = self._build_cartesian(
[…]
