### 2023-07-10 a956426
Triangle geometry * creating TriangleDomain * tests
