+++ pina/geometry/__init__.py
@@ -13,4 +13,8 @@ from .intersection_domain import Intersection
 from .union_domain import Union
 from .difference_domain import Difference
 from .union_domain import Union
+<<<<<<< HEAD
 from .triangle import SimplexDomain
+=======
+from .triangle import TriangleDomain
+>>>>>>> 053d570 (Triangle geometry)
