+++ pina/geometry/__init__.py
@@ -3,10 +3,7 @@ __all__ = [
     'CartesianDomain',
     'EllipsoidDomain',
     'Union',
-    'Intersection',
-    'Exclusion',
-    'Difference',
-    'OperationInterface'
+    'SimplexDomain'
 ]
 
 from .location import Location
@@ -16,4 +13,5 @@ from .exclusion_domain import Exclusion
 from .intersection_domain import Intersection
 from .union_domain import Union
 from .difference_domain import Difference
-from .operation_interface import OperationInterface
+from .union_domain import Union
+from .simplex import SimplexDomain
