+++ pina/geometry/simplex.py
@@ -1,0 +1,281 @@
+import torch
+from .location import Location
+from pina.geometry import CartesianDomain
+from pina import LabelTensor
+from ..utils import check_consistency
+
+class SimplexDomain(Location):
+    """PINA implementation of a Simplex."""
+
+    def __init__(self, simplex_matrix, sample_surface=False):
+        """
+        :param simplex_matrix: A matrix of LabelTensor objects representing
+            a vertex of the simplex (a tensor), and the coordinates of the
+            point (a list of labels).
+        :type simplex_matrix: list[LabelTensor]
+        :param sample_surface: A variable for choosing sample strategies. If
+            `sample_surface=True` only samples on the Simplex surface
+            frontier are taken. If `sample_surface=False`, no such criteria
+            is followed.
+        :type sample_surface: bool
+<<<<<<< HEAD
+=======
+
+>>>>>>> origin/v0.1-triangle
+        .. warning::
+            Sampling for dimensions greater or equal to 10 could result
+            in a shrinking of the simplex, which degrades the quality
+            of the samples. For dimensions higher than 10, other algorithms
+            for sampling should be used.
+<<<<<<< HEAD
+        :Example:
+=======
+
+        :Example:
+
+>>>>>>> origin/v0.1-triangle
+            >>> spatial_domain = SimplexDomain(
+                    [
+                        LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+                        LabelTensor(torch.tensor([[1, 1]]), labels=["x", "y"]),
+                        LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+                    ], sample_surface = True
+                )
+        """
+
+        # check consistency of sample_surface
+        check_consistency(sample_surface, bool)
+        self._sample_surface = sample_surface
+
+        # check consistency of simplex_dict
+        if not isinstance(simplex_matrix, list):
+            raise ValueError(f"{type(simplex_matrix).__name__} must be {list}.")
+        check_consistency(simplex_matrix, LabelTensor)
+
+        # check consistency of labels
+        if simplex_matrix[0].labels != simplex_matrix[1].labels:
+            raise ValueError(f"Labels don't match.")
+<<<<<<< HEAD
+
+        # check consistency of labels
+        if simplex_matrix[0].labels != simplex_matrix[1].labels:
+            raise ValueError(f"Labels don't match.")
+=======
+>>>>>>> origin/v0.1-triangle
+        
+        # vertices, vectors, coordinates
+        self._vertices_matrix = torch.tensor(
+            [
+                [float(vertex.extract(label)) for label in vertex.labels]
+                for vertex in simplex_matrix
+            ]
+        ).T
+        self._vectors_shifted = (
+            self._vertices_matrix[:, :-1] - self._vertices_matrix[:, None, -1]
+        )
+        self._coordinates = simplex_matrix[0].labels
+
+        # build cartesian_bound
+        self._cartesian_bound = self._build_cartesian(self.variables)
+
+    @property
+    def coordinates(self):
+        return self._coordinates
+
+    @property
+    def variables(self):
+        return self._vertices_matrix
+
+    def _build_cartesian(self, vertices):
+        """
+        Build Cartesian border for Simplex domain to be used in sampling.
+<<<<<<< HEAD
+=======
+
+>>>>>>> origin/v0.1-triangle
+        :param vertex_matrix: matrix of vertices
+        :type vertices: list[list]
+        :return: Cartesian border for triangular domain
+        :rtype: CartesianDomain
+        """
+
+        span_dict = {}
+
+        for i, coord in enumerate(self._coordinates):
+            sorted_vertices = sorted(vertices, key=lambda vertex: vertex[i])
+
[…]
