### 2023-07-10 7befa7a
reworked _vertices_matrix
