### 2023-07-10 c774a4c
removed float typecasting in vertices matrices
