+++ pina/geometry/union_domain.py
@@ -10,10 +10,16 @@ class Union(OperationInterface):
     """ PINA implementation of Unions of Domains."""
 
     def __init__(self, geometries):
-        """ PINA implementation of Unions of Domains.
+        """
+        PINA implementation of Unions of Domains.
+        Given two sets :math:`A` and :math:`B` then the
+        domain difference is defined as:
 
         ..:math:
-        A \cup B = \{x \mid x \in A \text{ or } x \in B\}
+        A \cup B = \{x \mid x \in A \text{ or } x \in B\},
+
+        with :math:`x` a point in :math:`\mathbb{R}^N` and :math:`N`
+        the dimension of the geometry space.
 
         :param list geometries: A list of geometries from 'pina.geometry' 
             such as 'EllipsoidDomain' or 'CartesianDomain'.
