+++ pina/geometry/difference_domain.py
@@ -9,10 +9,16 @@ class Difference(OperationInterface):
     """ PINA implementation of Difference of Domains."""
 
     def __init__(self, geometries):
-        """ PINA implementation of Difference of Domains.
+        """
+        PINA implementation of Difference of Domains.
+        Given two sets :math:`A` and :math:`B` then the
+        domain difference is defined as:
 
         ..:math:
-        A \setminus B = \{x \mid x \in A \text{ and } x \not\in B\}
+        A \setminus B = \{x \mid x \in A \text{ and } x \not\in B\},
+
+        with :math:`x` a point in :math:`\mathbb{R}^N` and :math:`N`
+        the dimension of the geometry space.
 
         :param list geometries: A list of geometries from 'pina.geometry' 
             such as 'EllipsoidDomain' or 'CartesianDomain'. The first 
