+++ pina/geometry/intersection_domain.py
@@ -10,10 +10,16 @@ class Intersection(OperationInterface):
     """ PINA implementation of Intersection of Domains."""
 
     def __init__(self, geometries):
-        """ PINA implementation of Intersection of Domains.
+        """
+        PINA implementation of Intersection of Domains.
+        Given two sets :math:`A` and :math:`B` then the
+        domain difference is defined as:
 
         ..:math:
-        A \cap B = \{x \mid x \in A \text{ and } x \in B\}
+        A \cap B = \{x \mid x \in A \text{ and } x \in B\},
+
+        with :math:`x` a point in :math:`\mathbb{R}^N` and :math:`N`
+        the dimension of the geometry space.
 
         :param list geometries: A list of geometries from 'pina.geometry' 
             such as 'EllipsoidDomain' or 'CartesianDomain'. The intersection
