+++ pina/geometry/difference_domain.py
@@ -9,7 +9,16 @@ class Difference(OperationInterface):
     """ PINA implementation of Difference of Domains."""
 
     def __init__(self, geometries):
-        """ PINA implementation of Difference of Domains.
+        """
+        PINA implementation of Difference of Domains.
+        Given two sets :math:`A` and :math:`B` then the
+        domain difference is defined as:
+
+        ..:math:
+        A \setminus B = \{x \mid x \in A \text{ and } x \not\in B\},
+
+        with :math:`x` a point in :math:`\mathbb{R}^N` and :math:`N`
+        the dimension of the geometry space.
 
         :param list geometries: A list of geometries from 'pina.geometry' 
             such as 'EllipsoidDomain' or 'CartesianDomain'. The first 
@@ -32,7 +41,7 @@ class Difference(OperationInterface):
             if geometry.is_inside(point):
                 return False
         return self.geometries[0].is_inside(point, check_border)
-    
+
     def sample(self, n, mode='random', variables='all'):
         """Sample routine for difference domain.
 
