### 2023-07-14 ea42e26
simple logger
