+++ pina/solvers/pinn.py
@@ -131,6 +131,10 @@ class PINN(SolverInterface):
         # TODO Fix the bug, tot_loss is a label tensor without labels
         # we need to pass it as a torch tensor to make everything work
         total_loss = sum(condition_losses)
+
+        self.log('mean_loss', float(total_loss / len(condition_losses)), prog_bar=True, logger=False)
+        for condition_loss, loss in zip(self.problem.conditions, condition_losses):
+            self.log(condition_loss + '_loss', float(loss), prog_bar=True, logger=False)
         return total_loss
 
     @property
