+++ tests/test_geometry/test_simplex.py
@@ -1,0 +1,141 @@
+import torch
+import pytest
+
+from pina import LabelTensor
+from pina.geometry import SimplexDomain
+
+
+def test_constructor():
+    SimplexDomain(
+        [
+            LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[1, 1]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+        ]
+    )
+    SimplexDomain(
+        [
+            LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[1, 1]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+        ],
+        sample_surface=True,
+    )
+    with pytest.raises(ValueError):
+        SimplexDomain(
+            [
+                LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+                LabelTensor(torch.tensor([[1, 1]]), labels=["x", "z"]),
+                LabelTensor(torch.tensor([[0, 2]]), labels=["x", "a"]),
+            ]
+        )
+        SimplexDomain(
+            [
+                LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+                [1, 1],
+                LabelTensor(torch.tensor([[0, 2]]), labels=["x", "y"]),
+            ]
+        )
+
+
+def test_is_inside_faulty_point():
+    domain = SimplexDomain(
+        [
+            LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[2, 2]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[4, 0]]), labels=["x", "y"]),
+        ],
+        sample_surface=True,
+    )
+    pt = LabelTensor(torch.tensor([[2, 1]]), ["x", "z"])
+    with pytest.raises(ValueError):
+        assert domain.is_inside(point=pt, check_border=False) == True
+
+
+def test_is_inside_2D_check_border_true():
+    domain = SimplexDomain(
+        [
+            LabelTensor(torch.tensor([[0, 0]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[2, 2]]), labels=["x", "y"]),
+            LabelTensor(torch.tensor([[4, 0]]), labels=["x", "y"]),
+        ],
+        sample_surface=True,
+    )
+    pt1 = LabelTensor(torch.tensor([[0, 0]]), ["x", "y"])
+    pt2 = LabelTensor(torch.tensor([[2, 2]]), ["x", "y"])
+    pt3 = LabelTensor(torch.tensor([[4, 0]]), ["x", "y"])
+    pt4 = LabelTensor(torch.tensor([[3, 1]]), ["x", "y"])
+    pt5 = LabelTensor(torch.tensor([[2, 1]]), ["x", "y"])
+    pt6 = LabelTensor(torch.tensor([[100, 100]]), ["x", "y"])
+    pts = [pt1, pt2, pt3, pt4, pt5, pt6]
+    for pt, exp_result in zip(pts, [True, True, True, True, False, False]):
+        assert domain.is_inside(point=pt, check_border=True) == exp_result
+
+
+def test_is_inside_2D_check_border_false():
[…]
