### 2023-07-14 ef7eeb6
Simplex Domain (#152)
* add Simplex Domain 
* unittests for Simplex
* docs update

---------

Co-authored-by: Dario Coscia <93731561+dario-coscia@users.noreply.github.com>
