### 2023-07-17 cf87c63
Basic TensorBoard Implementation - test_loggers contains a basic script to test the loggers     - run this to get data to be able to plot and see on tensorboard - changed trainer to take in a logger (tensorbaord)     - this should allow you to log stuff - used Dario's self.log code in pinn to log stuff
