+++ README.md
@@ -1,64 +1,171 @@
-<p align="center">
-    <a href="http://mathlab.github.io/PINA/" target="_blank" >
-    <img alt="Physics-Informed Neural networks for Advanced modeling" src="readme/pina_logo.png" width="200" />
+<!-- PROJECT SHIELDS -->
+<!--
+*** I'm using markdown "reference style" links for readability.
+*** Reference links are enclosed in brackets [ ] instead of parentheses ( ).
+*** See the bottom of this document for the declaration of the reference variables
+*** for contributors-url, forks-url, etc. This is an optional, concise syntax you may use.
+*** https://www.markdownguide.org/basic-syntax/#reference-style-links
+-->
+
+<div align="center">
+
+[![PyPi][pypi-shield]][pypi-url]
+[![PyPi][pypiversion-shield]][pypi-url]
+[![PyPi][downloads-shield]][downloads-url]
+[![License][license-shield]][license-url]
+
+</div>
+<!-- [![Contributors][contributors-shield]][contributors-url] -->
+<!-- [![Build][build-shield]][build-url] -->
+<!-- [![CodeCov][codecov-shield]][codecov-url] -->
+
+<!-- MARKDOWN LINKS & IMAGES -->
+<!-- https://www.markdownguide.org/basic-syntax/#reference-style-links -->
+
+[pypi-shield]: https://img.shields.io/pypi/pyversions/pina-mathlab?style=for-the-badge
+
+[pypi-url]: https://pypi.org/project/pina-mathlab/
+
+[pypiversion-shield]: https://img.shields.io/pypi/v/pina-mathlab?style=for-the-badge
+
+[downloads-shield]: https://img.shields.io/pypi/dm/pina-mathlab?style=for-the-badge
+
+[downloads-url]: https://pypi.org/project/zenml/
+
+[codecov-shield]: https://img.shields.io/codecov/c/gh/zenml-io/zenml?style=for-the-badge
+
+[codecov-url]: https://codecov.io/gh/zenml-io/zenml
+
+[contributors-shield]: https://img.shields.io/github/contributors/zenml-io/zenml?style=for-the-badge
+
+[contributors-url]: https://github.com/othneildrew/Best-README-Template/graphs/contributors
+
+[license-shield]: https://img.shields.io/github/license/mathLab/pina?style=for-the-badge
+
+[license-url]: https://github.com/mathLab/PINA/blob/main/LICENSE.rst
+
+[linkedin-shield]: https://img.shields.io/badge/-LinkedIn-black.svg?style=for-the-badge&logo=linkedin&colorB=555
+
+<!-- [linkedin-url]: https://www.linkedin.com/company/zenml/
+
+[twitter-shield]: https://img.shields.io/twitter/follow/zenml_io?style=for-the-badge
+
+[twitter-url]: https://twitter.com/zenml_io -->
+
+[slack-shield]: https://img.shields.io/badge/-Slack-black.svg?style=for-the-badge&logo=linkedin&colorB=555
+
+[slack-url]: https://zenml.io/slack-invite
+
+[build-shield]: https://img.shields.io/github/workflow/status/zenml-io/zenml/Build,%20Lint,%20Unit%20&%20Integration%20Test/develop?logo=github&style=for-the-badge
+
+[build-url]: https://github.com/zenml-io/zenml/actions/workflows/ci.yml
+
+<!-- PROJECT LOGO -->
+<br />
+<div align="center">
+  <a href="https://zenml.io">
+    <img alt="ZenML Logo" src="https://github.com/mathLab/PINA/raw/master/readme/pina_logo.png" alt="Logo" width="200">
+  </a>
+
+<h3 align="center">Solve equations, intuitively.</h3>
+
+  <p align="center">
[…]
