### 2023-07-17 e2d7920
Create CONTRIBUTING.md
