### 2023-07-18 b25c510
device training
