+++ pina/trainer.py
@@ -4,6 +4,8 @@ import lightning.pytorch as pl
 from .utils import check_consistency
 from .dataset import DummyLoader
 from .solvers.solver import SolverInterface
+from pytorch_lightning.loggers import TensorBoardLogger
+
 
 class Trainer(pl.Trainer):
 
@@ -28,7 +30,19 @@ class Trainer(pl.Trainer):
         # TODO: make a better dataloader for train
         self._loader = DummyLoader(solver.problem.input_pts, device) 
 
+        # handling loggers
+        self._pina_logger = PinaLogger(self.loggers)
 
-    def train(self): # TODO add kwargs and lightining capabilities
-        return super().fit(self._model, self._loader)
-    
+        # self._logger = self._logger_connector.trainer.loggers
+
+        # Initialize TensorBoardLogger
+        self.logger = TensorBoardLogger('logs/', name='my_experiment')
+
+    def train(self):  # TODO add kwargs and lightining capabilities
+        self.logger.log_hyperparams(self._model.hparams)
+        self.logger.log_graph(self._model)
+
+        super().fit(self._model, self._loader)
+
+        # self.logger.experiment.close()
+        # self.close()
