### 2023-07-18 d19faca
Refactoring Logger: - moved self.log code into pina/solver.py - created a logger class with basic functionality and checks - edited trainer to use Pina Logger and log_graph - removed logging from pinn.py
