+++ pina/trainer.py
@@ -4,7 +4,7 @@ import lightning.pytorch as pl
 from .utils import check_consistency
 from .dataset import DummyLoader
 from .solvers.solver import SolverInterface
-from pytorch_lightning.loggers import TensorBoardLogger
+from .loggers import PinaLogger
 
 
 class Trainer(pl.Trainer):
@@ -30,14 +30,19 @@ class Trainer(pl.Trainer):
         # TODO: make a better dataloader for train
         self._loader = DummyLoader(solver.problem.input_pts, device) 
 
+        # handling loggers
+        self._pina_logger = PinaLogger(self.loggers)
+
+        # self._logger = self._logger_connector.trainer.loggers
+
         # Initialize TensorBoardLogger
-        self.logger = TensorBoardLogger('logs/', name='my_experiment')
+        # self.logger = TensorBoardLogger('logs/', name='my_experiment')
 
     def train(self):  # TODO add kwargs and lightining capabilities
-        self.logger.log_hyperparams(self._model.hparams)
-        self.logger.log_graph(self._model)
+        # self.logger.log_hyperparams(self._model.hparams)
+        # print(self._model.hparams)
+        # self.loggers[0].log_graph(model=self._model)
+
+        self._pina_logger.pina_log_graph(model=self._model)
 
         super().fit(self._model, self._loader)
-
-        # self.logger.experiment.close()
-        # self.close()
