### 2023-07-18 ac250e7
Cleaned up code
