+++ pina/trainer.py
@@ -30,19 +30,11 @@ class Trainer(pl.Trainer):
         # TODO: make a better dataloader for train
         self._loader = DummyLoader(solver.problem.input_pts, device) 
 
-        # handling loggers
+        # Initalize logger using PinaLogger
         self._pina_logger = PinaLogger(self.loggers)
-
-        # self._logger = self._logger_connector.trainer.loggers
-
-        # Initialize TensorBoardLogger
-        # self.logger = TensorBoardLogger('logs/', name='my_experiment')
+        # self._pina_logger.pina_log_graph(model=self._model)
 
     def train(self):  # TODO add kwargs and lightining capabilities
-        # self.logger.log_hyperparams(self._model.hparams)
-        # print(self._model.hparams)
-        # self.loggers[0].log_graph(model=self._model)
-
         self._pina_logger.pina_log_graph(model=self._model)
 
         super().fit(self._model, self._loader)
