### 2023-07-18 593a860
Cleaned up code
