### 2023-07-19 #5352 @ndem0
Dear @danielskatz, I should have done all the points!

- the tag is `0.0.3-joss`, here the link to the release https://github.com/mathLab/PINA/releases/tag/v0.0.3-joss
- I uploaded that tag to Zenodo, this is the record page https://zenodo.org/record/8163732
- the DOI for the Zenodo capsule is https://doi.org/10.5281/zenodo.8163732

I double checked and everything looks correct to me, but let me know if there are some problem and I'll fix asap! Thanks
