+++ tests/test_solvers/test_pinn.py
@@ -121,14 +121,15 @@ def test_train_extra_feats_cpu():
     trainer = Trainer(solver=pinn, kwargs={'max_epochs' : 5, 'accelerator':'cpu'})
     trainer.train()
 
-def test_train_gpu(): 
-    poisson_problem = Poisson()
-    boundaries = ['gamma1', 'gamma2', 'gamma3', 'gamma4']
-    n = 10
-    poisson_problem.discretise_domain(n, 'grid', locations=boundaries)
-    pinn = PINN(problem = poisson_problem, model=model, extra_features=None, loss=LpLoss())
-    trainer = Trainer(solver=pinn, kwargs={'max_epochs' : 5, 'accelerator':'gpu'})
-    trainer.train()
+# TODO, fix GitHub actions to run also on GPU
+# def test_train_gpu(): 
+#     poisson_problem = Poisson()
+#     boundaries = ['gamma1', 'gamma2', 'gamma3', 'gamma4']
+#     n = 10
+#     poisson_problem.discretise_domain(n, 'grid', locations=boundaries)
+#     pinn = PINN(problem = poisson_problem, model=model, extra_features=None, loss=LpLoss())
+#     trainer = Trainer(solver=pinn, kwargs={'max_epochs' : 5, 'accelerator':'gpu'})
+#     trainer.train()
 
 """
 def test_train_gpu(): #TODO fix ASAP
