### 2023-07-19 #5352 @ndem0
Yes there were some errors, now the authors string is fine
