+++ joss/paper.bib
@@ -65,7 +65,7 @@ abstract = {We introduce physics-informed neural networks – neural networks th
 
 %%other PINN packages
 @article{chen2020neurodiffeq,
-  title={Neurodiffeq: A python package for solving differential equations with neural networks},
+  title={Neurodiffeq: A {P}ython package for solving differential equations with neural networks},
   author={Chen, Feiyu and Sondak, David and Protopapas, Pavlos and Mattheakis, Marios and Liu, Shuheng and Agarwal, Devansh and Di Giovanni, Marco},
   journal={Journal of Open Source Software},
   doi = {10.21105/joss.01931},
@@ -89,12 +89,14 @@ abstract = {We introduce physics-informed neural networks – neural networks th
   title={TensorDiffEq: Scalable Multi-GPU Forward and Inverse Solvers for Physics Informed Neural Networks},
   author={McClenny, Levi D and Haile, Mulugeta A and Braga-Neto, Ulisses M},
   journal={arXiv preprint arXiv:2103.16034},
+  doi={10.48550/arXiv.2103.16034},
   year={2021}
 }
 @article{peng2021idrlnet,
   title={IDRLnet: A physics-informed neural network library},
   author={Peng, Wei and Zhang, Jun and Zhou, Weien and Zhao, Xiaoyu and Yao, Wen and Chen, Xiaoqian},
   journal={arXiv preprint arXiv:2107.04320},
+  doi={10.48550/arXiv.2107.04320},
   year={2021}
 }
 @inproceedings{hennigh2021nvidia,
@@ -107,7 +109,7 @@ abstract = {We introduce physics-informed neural networks – neural networks th
   organization={Springer}
 }
 @article{haghighat2021sciann,
-  title={Sciann: A keras/tensorflow wrapper for scientific computations and physics-informed deep learning using artificial neural networks},
+  title={Sciann: A {K}eras/{T}ensor{F}low wrapper for scientific computations and physics-informed deep learning using artificial neural networks},
   author={Haghighat, Ehsan and Juanes, Ruben},
   journal={Computer Methods in Applied Mechanics and Engineering},
   doi = {10.1016/j.cma.2020.113552},
@@ -117,15 +119,17 @@ abstract = {We introduce physics-informed neural networks – neural networks th
   publisher={Elsevier}
 }
 @article{koryagin2019pydens,
-  title={PyDEns: A python framework for solving differential equations with neural networks},
+  title={PyDEns: A {P}ython framework for solving differential equations with neural networks},
   author={Koryagin, Alexander and Khudorozkov, Roman and Tsimfer, Sergey},
   journal={arXiv preprint arXiv:1909.11544},
+  doi={10.48550/arXiv.1909.11544},
   year={2019}
 }
 @article{araz2021elvet,
-  title={Elvet--a neural network-based differential equation and variational problem solver},
+  title={Elvet -- a neural network-based differential equation and variational problem solver},
   author={Araz, Jack Y and Criado, Juan Carlos and Spannowsky, Michael},
   journal={arXiv preprint arXiv:2103.14575},
+  doi={10.48550/arXiv.2103.14575},
   year={2021}
 }
 
