### 2023-07-19 #5352 @danielskatz
@ndem0 - I'm suggesting some minor changes in https://github.com/mathLab/PINA/pull/161 - please merge this, or let me know what you disagree with, then we can proceed to publication
